% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routes.R
\name{api_pat}
\alias{api_pat}
\title{Retrieve personal access token.}
\usage{
api_pat(api_name, force = FALSE)
}
\arguments{
\item{api_name}{Text string of the name of the API you are calling, e.g.
cyclestreet, graphhopper etc.}
}
\description{
Retrieve personal access token.
}
\examples{
\dontrun{
api_pat(api_name = "cyclestreet")
}
}
\keyword{internal}

