% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo-functions.R
\name{bb2poly}
\alias{bb2poly}
\title{Convert a bounding box to a SpatialPolygonsDataFrame}
\usage{
bb2poly(bb)
}
\arguments{
\item{bb}{the bounding box or spatial object that will be used to crop \code{shp}}
}
\description{
Takes a bounding box as an input and outputs a box in the form of a polygon
}
\examples{
bb <- structure(c(-1.55080650299106, 53.8040984493515, -1.51186138683098,
53.828874094091), .Dim = c(2L, 2L), .Dimnames = list(c("coords.x1",
  "coords.x2"), c("min", "max")))
bb1 <- bb2poly(bb)
plot(bb1)
}
