% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linefuns.R
\name{bind_sf}
\alias{bind_sf}
\title{Rapid row-binding of sf objects}
\usage{
bind_sf(x)
}
\arguments{
\item{x}{List of sf objects to combine}
}
\value{
An sf data frame
}
\description{
Rapid row-binding of sf objects
}
