% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ca_local}
\alias{ca_local}
\title{SpatialPointsDataFrame representing road traffic deaths}
\format{
A SpatialPointsDataFrame with 11 rows and 2 columns
}
\usage{
data(ca_local)
}
\description{
This dataset represents the type of data downloaded and cleaned
using stplanr functions. It represents a very small sample (with most variables stripped)
of open data from the UK's Stats19 dataset.
}
\keyword{datasets}
