% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialLinesNetwork.R
\name{calc_network_routes}
\alias{calc_network_routes}
\title{Find shortest path between nodes on network}
\usage{
calc_network_routes(sln, start, end, sumvars)
}
\arguments{
\item{sln}{The SpatialLinesNetwork to use.}

\item{start}{Node ID(s) of route starts.}

\item{end}{Node ID(s) of route ends.}

\item{sumvars}{Character vector of variables for which to calculate
summary statistics.}
}
\description{
Find shortest path between nodes on network
}
\section{Details}{

Find the shortest path on the network between specified nodes and returns
a SpatialLinesdataFrame containing the path(s) and summary statistics of the
route.
}

