% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cents_sf}
\alias{cents_sf}
\title{Spatial points representing home locations}
\format{
A spatial dataset with 8 rows and 5 columns
}
\description{
These points represent population-weighted centroids of Medium Super Output Area (MSOA) zones within a 1 mile radius of of my home when I was writing this package.
}
\details{
\itemize{
\item geo_code the official code of the zone
\item MSOA11NM name zone name
\item percent_fem the percent female
\item avslope average gradient of the zone
}

Cents was generated from the data repository pct-data: https://github.com/npct/pct-data. This data was accessed from within the pct repo: https://github.com/npct/pct, using the following code:
}
\examples{
cents_sf
}
\seealso{
Other data: 
\code{\link{destinations_sf}},
\code{\link{flow}},
\code{\link{flow_dests}},
\code{\link{flowlines_sf}},
\code{\link{od_data_lines}},
\code{\link{od_data_routes}},
\code{\link{od_data_sample}},
\code{\link{osm_net_example}},
\code{\link{read_table_builder}()},
\code{\link{route_network_sf}},
\code{\link{route_network_small}},
\code{\link{routes_fast_sf}},
\code{\link{routes_slow_sf}},
\code{\link{zones_sf}}
}
\concept{data}
\keyword{datasets}
