% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/distance-decay-functions.R
\name{dd_iac}
\alias{dd_iac}
\title{log-linear distance-decay function}
\usage{
dd_iac(d, a = 0.3, b = 0.2)
}
\description{
log-linear distance-decay function
}
\section{Details}{

Implementation of the formula used to fit distance to probability of
 travelling by a given mode by Iacono et al. (2010)
}
\examples{
d <- 0:10 # vector of distances
params <- c(0, -0.002, -0.15, -3)
(res <- dd_iac(d, c(0, -0.002, -0.15, -3)))
plot(d, res)

d = seq(0, 50, 0.1)
plot(d, dd_iac(d))
lines(d, dd_iac(d, b = 0.15))
}
\references{
Iacono, M., Krizek, K. J., & El-Geneidy, A. (2010).
Measuring non-motorized accessibility: issues, alternatives, and execution. Journal of Transport Geography, 18(1), 133–140. doi:10.1016/j.jtrangeo.2009.02.002
}
\seealso{
\code{\link{dd_logsqr}}
}

