% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/distance-decay-functions.R
\name{dd_logcub}
\alias{dd_logcub}
\title{Cubic logarithmic decay function}
\usage{
dd_logcub(d, par)
}
\description{
Cubic logarithmic decay function
}
\section{Details}{

Function for converting distance of a trip into the probability of travel by
a particular mode.
}
\examples{
d <- 0:10 # vector of distances
params <- c(0, -0.002, -0.15, -3)
(res <- dd_logcub(d, c(0, -0.002, -0.15, -3)))
plot(d, res)

par_male_urb <- c(0.0001945, -0.002273, -0.1489583, -2.923221)
par_female_urb <- c(-0.0017932, 0.0655261, -0.8396819, -2.892149)
d = seq(0, 50, 0.1)
plot(d, dd_logcub(d, par = par_male_urb), ylim = c(0, 0.1)) # test plots
lines(d, dd_logcub(d, par = par_female_urb), ylim = c(0, 0.1)) # test plots
}

