% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/distance-decay-functions.R
\name{dd_logsqr}
\alias{dd_logsqr}
\title{log-linear-square-root distance decay}
\usage{
dd_logsqr(d, a, b, c = 0)
}
\description{
log-linear-square-root distance decay
}
\section{Details}{

log-linear distance decay (Iacono et al., 2010) with an additional
square-root term to make the curve much more flexible. The additional
term (c in the function below) describes the short-term response:
positive c values result in a unimodal distance decay with a peak after d = 0;
negative c values lead to rapid decay in cycling under 2 km.
}
\examples{
d = seq(0, 50, 0.1)
plot(d, dd_logsqr(d, a = 0.3, b = -0.2, c = -0.5), ylim = c(0, 0.5))
lines(d, dd_logsqr(d, 0.3, -0.2, 0.5))
}
\references{
Iacono, M., Krizek, K. J., & El-Geneidy, A. (2010).
Measuring non-motorized accessibility: issues, alternatives, and execution. Journal of Transport Geography, 18(1), 133–140. doi:10.1016/j.jtrangeo.2009.02.002
}

