% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/osrmfunctions.R
\name{decode_google_lines}
\alias{decode_google_lines}
\title{Decode Google polyline compressed string}
\usage{
decode_google_lines(polyline)
}
\arguments{
\item{polyline}{A character string or vector of character strings containing
the encoded polyline to be decoded.}
}
\description{
Decode Google polyline compressed string
}
\section{Details}{

An implementation of the Google Maps Encoded Polyline Algorithm for decoding
strings. Returns a dataframe if polyline is of length 1 and a list of
dataframes otherwise.
}
\examples{
\dontrun{
 decode_google_lines('')
}
}

