% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{destination_zones}
\alias{destination_zones}
\alias{destinations}
\alias{destinations_sf}
\title{Example destinations data}
\format{
A spatial dataset with 87 features
}
\usage{
data(destination_zones)
}
\description{
This dataset represents trip destinations on a different geographic
level than the origins stored in the object \code{cents}.
}
\details{
See https://github.com/ropensci/stplanr/blob/v0.8.5/R/data.R for
details on how the datasets were created.
}
\examples{
destinations_sf
}
\seealso{
Other example data: 
\code{\link{flow_dests}},
\code{\link{flowlines}},
\code{\link{flow}},
\code{\link{route_network}},
\code{\link{routes_fast}},
\code{\link{routes_slow}}
}
\concept{example data}
\keyword{datasets}
