% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google-functions.R
\name{dist_google}
\alias{dist_google}
\title{Return travel network distances and time using the Google Maps API}
\usage{
dist_google(lat, lng, google_api = "", units = "metric")
}
\arguments{
\item{lat}{Numeric vector containing latitude coordinate for each coordinate
to map. Also accepts dataframe with latitude in the first column and
longitude in the second column.}

\item{lng}{Numeric vector containing longitude coordinate for each
coordinate to map.}

\item{google_api}{String value containing the Google API key to use.}

\item{units}{Text string, either metric (default) or imperial.}
}
\description{
Return travel network distances and time using the Google Maps API
}
\section{Details}{

Estimate travel times accounting for the road network - see \url{https://developers.google.com/maps/documentation/distance-matrix/}
Note: Currently returns the json object returned by the Google Maps API and uses the same origins and destinations.
}
\examples{
\dontrun{
 nearest_google(lat = 50.333, lng = 3.222, google_api = "api_key_here")
}
}

