% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-stats19.R
\name{dl_stats19}
\alias{dl_stats19}
\title{Download Stats19 data}
\usage{
dl_stats19(zip_url = paste0("http://data.dft.gov.uk.s3.amazonaws.com/",
  "road-accidents-safety-data/Stats19_Data_2005-2014.zip"))
}
\arguments{
\item{zip_url}{The url where the data is stored}
}
\description{
Download Stats19 data
}
\section{Details}{

This convenience function downloads and unzips UK road traffic casualty data.
It results in unzipped .csv data in R's temporary directory.

Ensure you have a fast internet connection and at least 100 Mb space
}
\examples{
\dontrun{
dl_stats19()
}
}

