% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{flowlines}
\alias{flowlines}
\title{SpatialLinesDataFrame of commuter flows}
\format{A SpatialLinesDataFrame 49 rows and 15 columns}
\usage{
data(flowlines)
}
\description{
SpatialLinesDataFrame of commuter flows
}
\keyword{datasets}

