% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flowlines_sf}
\alias{flowlines_sf}
\title{Spatial lines dataset of commuter flows}
\format{
A spatial lines dataset with 42 rows and 15 columns
}
\description{
Flow data after conversion to a spatial format..
}
\seealso{
Other data: 
\code{\link{cents_sf}},
\code{\link{destinations_sf}},
\code{\link{flow}},
\code{\link{flow_dests}},
\code{\link{od_data_lines}},
\code{\link{od_data_routes}},
\code{\link{od_data_sample}},
\code{\link{osm_net_example}},
\code{\link{read_table_builder}()},
\code{\link{route_network_sf}},
\code{\link{route_network_small}},
\code{\link{routes_fast_sf}},
\code{\link{routes_slow_sf}},
\code{\link{zones_sf}}
}
\concept{data}
\keyword{datasets}
