% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-stats19.R
\name{format_stats19_ve}
\alias{format_stats19_ve}
\title{Format UK 'Stats19' road traffic casualty data}
\usage{
format_stats19_ve(ve)
}
\arguments{
\item{ve}{Dataframe representing the raw Stats19 data read-in with \code{read_csv()}.}
}
\description{
Format UK 'Stats19' road traffic casualty data
}
\section{Details}{

This is a helper function to format raw stats19 data
}
\examples{
\dontrun{
ve <- format_stats19_ve(ve)
}
}

