% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo-functions.R
\name{geo_bb_matrix}
\alias{geo_bb_matrix}
\title{Create matrix representing the spatial bounds of an object}
\usage{
geo_bb_matrix(shp)
}
\arguments{
\item{shp}{Spatial object}
}
\description{
Converts a range of spatial data formats into a matrix representing the bounding box
}
\examples{
geo_bb_matrix(routes_fast_sf)
geo_bb_matrix(cents_sf[1, ])
geo_bb_matrix(c(-2, 54))
geo_bb_matrix(sf::st_coordinates(cents_sf))
}
\seealso{
Other geo: 
\code{\link{bbox_scale}()},
\code{\link{geo_bb}()},
\code{\link{quadrant}()}
}
\concept{geo}
