% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_code.R
\name{geo_code}
\alias{geo_code}
\title{Convert text strings into points on the map}
\usage{
geo_code(address, base_url = "http://maps.google.com", sensor = "false")
}
\arguments{
\item{address}{Text string representing the address you want to geocode}

\item{base_url}{The base url to query}
}
\description{
Generate a lat/long pair from data using Google's geolocation API.
}
\examples{
address = "LS7 3HB"
geo_code(address = address)
}
