% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_projected.R
\name{geo_select_aeq}
\alias{geo_select_aeq}
\title{Select a projected CRS}
\usage{
geo_select_aeq(shp)
}
\description{
Select a projected CRS
}
\examples{
shp = st_sf(st_sfc(st_point(c(1, 0))))
geo_select_aeq(shp)
}
