% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gFlow.R
\name{gFlow2line}
\alias{gFlow2line}
\title{Convert flow data to SpatialLinesDataFrame}
\usage{
gFlow2line(flow, zones)
}
\description{
Convert flow data to SpatialLinesDataFrame
}
\section{Details}{

Origin-destination ('OD') flow data is often provided
in the form of 1 line per flow with zone codes of origin and destination
centroids. This can be tricky to plot and link-up with geographical data.
This function makes the task easier.
}
\references{
Rae, A. (2009). From spatial interaction data to spatial interaction information? Geovisualisation and spatial structures of migration from the 2001 UK census. Computers, Environment and Urban Systems, 33(3), 161–178. doi:10.1016/j.compenvurbsys.2009.01.007
}

