% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gFlow.R
\name{gLines2CyclePath}
\alias{gLines2CyclePath}
\title{Convert straight SpatialLinesDataFrame from flow data in routes}
\usage{
gLines2CyclePath(lines, plan = "fastest")
}
\arguments{
\item{lines}{A SpatialLinesDataFrame object}

\item{plan}{A text string. Must be either "balanced", "fastest" (default)
or "quietest"}
}
\description{
Convert straight SpatialLinesDataFrame from flow data in routes
}
\section{Details}{


You will need to have a cyclestreets api key for this function to work

mytoken <- readLines("~/Dropbox/dotfiles/cyclestreets-api-key-rl")

Sys.setenv(CYCLESTREET = mytoken) # see http://www.cyclestreets.net/api/

cckey <- Sys.getenv('CYCLESTREET')
}

