% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo-functions.R
\name{gprojected}
\alias{gprojected}
\title{Return information about a spatial object using a temporary, projected CRS}
\usage{
gprojected(sp_obj, fun = rgeos::gLength, new_proj = crs_select_aeq(sp_obj),
  ...)
}
\arguments{
\item{sp_obj}{A spatial object with a geographic CRS (WGS84)
around which a buffer should be drawn}

\item{fun}{A geographic function such as rgeos::gArea or rgeos::gLength (the default)}

\item{new_proj}{The new projection to use - by default selected by
\code{\link{crs_select_aeq}}}

\item{...}{Additional arguments passed to the function \code{fun}.}
}
\description{
This function performs geographic queries on objects not in their native CRS
(which is assumed to be lat/lon, or geographic) but in a projected CRS.
}
\examples{
sp_obj = flowlines
proj4string(sp_obj) = CRS("+init=epsg:4326")
gprojected(sp_obj) # the total length of of the lines
gprojected(sp_obj, byid = TRUE)
buff = buff_geo(sp_obj, width = 25)
gprojected(buff, fun = rgeos::gArea) # area in m2
}

