% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfs.R
\name{gtfs2sldf}
\alias{gtfs2sldf}
\title{Import GTFS shapes and route data to SpatialLinesDataFrame.}
\usage{
gtfs2sldf(gtfszip = "")
}
\arguments{
\item{gtfszip}{String with the file path of the GTFS feed zip file}
}
\description{
Takes a string with the file path of the zip file with the GTFS feed,
imports the shapes (geometry), route and agency data and returns a
SpatialLinesDataFrame for the GTFS feed.
}

