% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/distance-decay-functions.R
\name{iac}
\alias{iac}
\title{Estimate propensity to cycle}
\usage{
iac(x, a = 0.3, b = 0.2)
}
\description{
Distance-decay function (Iacono et al. 2011)
}
\section{Details}{

Implementation of the formula used to fit distance to probability of
 travelling by a given mode by Iacono et al. (2010)
}
\references{
Iacono, M., Krizek, K. J., & El-Geneidy, A. (2010).
Measuring non-motorized accessibility: issues, alternatives, and execution. Journal of Transport Geography, 18(1), 133–140. doi:10.1016/j.jtrangeo.2009.02.002
}

