% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overline_sf.R
\name{is_linestring}
\alias{is_linestring}
\title{Identify overlapping lines}
\usage{
is_linestring(sl, geom_type = "LINESTRING")
}
\arguments{
\item{sl}{A SpatialLinesDataFrame with overlapping elements}

\item{geom_type}{The type of geometry to check for, \code{LINESTRING} by default}
}
\description{
A small wrapper around \code{sf::st_geometry_type()} that
identifies lines that are \code{LINESTRING}s
}
