% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overline.R
\name{islines}
\alias{islines}
\title{Do the intersections between two geometries create lines?}
\usage{
islines(g1, g2)
}
\arguments{
\item{g1}{A spatial object}

\item{g2}{A spatial object}
}
\description{
This is a function required in \code{\link[=overline]{overline()}}. It identifies
whether sets of lines overlap (beyond shared points) or
not.
}
\examples{
\dontrun{
# sf implementation
islines(routes_fast_sf[2, ], routes_fast_sf[3, ])
islines(routes_fast_sf[2, ], routes_fast_sf[22, ])
}
}
\seealso{
Other rnet: 
\code{\link{gsection}()},
\code{\link{overline}()},
\code{\link{rnet_breakup_vertices}()},
\code{\link{rnet_group}()}
}
\concept{rnet}
