% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{l_poly}
\alias{l_poly}
\title{Line polygon}
\format{
A SpatialPolygon
}
\usage{
data(l_poly)
}
\description{
This dataset represents road width for testing.
}
\examples{
l_poly
}
\keyword{datasets}
