% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/od-funs.R
\name{line2df}
\alias{line2df}
\title{Convert straight SpatialLinesDataFrame to a data.frame with from and to coords}
\usage{
line2df(l)
}
\arguments{
\item{l}{A SpatialLinesDataFrame}
}
\description{
Convert straight SpatialLinesDataFrame to a data.frame with from and to coords
}
\examples{
data(flowlines) # load demo flowlines dataset
ldf <- line2df(flowlines)
}

