% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linefuns.R
\name{line_bearing}
\alias{line_bearing}
\title{Find the bearing of straight lines}
\usage{
line_bearing(l)
}
\arguments{
\item{l}{A SpatialLinesDataFrame}
}
\description{
This is a simple wrapper around the geosphere function \code{\link{bearing}} to return the
bearing (in degrees relative to north) of lines
}
\details{
Returns a boolean vector. TRUE means that the associated line is in fact a point
(has no distance). This can be useful for removing data that will not be plotted.
}
\examples{
line_bearing(flowlines)
}

