% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnet_join.R
\name{line_cast}
\alias{line_cast}
\title{Convert multilinestring object into linestrings, without losing and vertices}
\usage{
line_cast(x)
}
\arguments{
\item{x}{LInestring object}
}
\description{
Convert multilinestring object into linestrings, without losing and vertices
}
