% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linefuns.R
\name{line_length}
\alias{line_length}
\title{Calculate length of lines in geographic CRS}
\usage{
line_length(l, byid = TRUE)
}
\arguments{
\item{l}{A SpatialLinesDataFrame}

\item{byid}{Logical determining whether the length is returned per object (default is true)}
}
\description{
Calculate length of lines in geographic CRS
}
