% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overline-sf.R
\name{line_section}
\alias{line_section}
\title{Disag}
\usage{
line_section(sl)
}
\description{
Disag
}
\examples{
sl <- routes_fast
rsec <- line_section(routes_fast)
plot(routes_fast)
lines(rsec, col = "red", lwd = 3)
length(routes_fast)
length(rsec)
set.seed(5)
sel <- sample(length(rsec), 20)
plot(rsec[sel,], col = "blue", add = TRUE, lwd = 3) # overlapping lines
}
