% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/distance-decay-functions.R
\name{log_cubic}
\alias{log_cubic}
\title{Cubic logarithmic decay function}
\usage{
log_cubic(d, par)
}
\description{
Cubic logarithmic decay function
}
\section{Details}{

Function for converting distance of a trip into the probability of travel by a particular mode.
}

