% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geo-functions.R
\name{mapshape}
\alias{mapshape}
\title{Simplify geometry file of a shapfile.}
\usage{
mapshape(dsn, percent)
}
\arguments{
\item{dsn}{A character string providing the absolute path to the shapefile to simplify.}

\item{percent}{A number between 0 and 100 stating how aggressively to simplify
 the object
Percentage of removable points to retain.
So \code{percent = 1} is a very aggressive simplication, saving a huge amount of
hard-disk space.}
}
\description{
Simplify geometry file of a shapfile.
}
\section{Details}{

This is a wrapper funtion for the open source JavaScript command-line GIS application mapshaper: \url{https://github.com/mbloch/mapshaper} . mapshaper which must be installed locally for mapshaper to work. Writes \code{mapshape} writes new file to disk. Thanks to Richard and Adrian Ellison for demonstrating this in R.
}
\examples{
\dontrun{
mapshape("~/geodata/myShapefile.shp", 5)
}
}

