% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line_via.R
\name{mats2line}
\alias{mats2line}
\title{Convert 2 matrices to lines}
\usage{
mats2line(mat1, mat2)
}
\arguments{
\item{mat1}{Matrix representing origins}

\item{mat2}{Matrix representing origins}
}
\description{
Convert 2 matrices to lines
}
\examples{
{
m1 <- matrix(c(1, 2, 1, 2), ncol = 2)
m2 <- matrix(c(9, 9, 9, 1), ncol = 2)
l <- mats2line(m1, m2)
class(l)
lsf <- sf::st_sf(l, crs = 4326)
class(lsf)
plot(lsf)
# mapview::mapview(lsf)
}
}
