% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linefuns.R
\name{n_segments}
\alias{n_segments}
\title{Vectorised function to calculate number of segments given a max segment length}
\usage{
n_segments(line_length, max_segment_length)
}
\arguments{
\item{line_length}{The length of the line}

\item{max_segment_length}{The maximum length of each segment}
}
\description{
Vectorised function to calculate number of segments given a max segment length
}
\examples{
n_segments(50, 10)
n_segments(50.1, 10)
n_segments(1, 10)
n_segments(1:9, 2)
}
\seealso{
Other lines: 
\code{\link{angle_diff}()},
\code{\link{geo_toptail}()},
\code{\link{is_linepoint}()},
\code{\link{line2df}()},
\code{\link{line2points}()},
\code{\link{line_bearing}()},
\code{\link{line_breakup}()},
\code{\link{line_midpoint}()},
\code{\link{line_segment}()},
\code{\link{line_segment1}()},
\code{\link{line_via}()},
\code{\link{mats2line}()},
\code{\link{n_vertices}()},
\code{\link{onewaygeo}()},
\code{\link{points2line}()},
\code{\link{toptail_buff}()}
}
\concept{lines}
