% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyclestreets.R
\name{nearest_cyclestreets}
\alias{nearest_cyclestreets}
\title{Generate nearest point on the route network of a point using the CycleStreets.net}
\usage{
nearest_cyclestreets(lat, lng, pat = cyclestreet_pat())
}
\arguments{
\item{lat}{Numeric vector containing latitude coordinate for each coordinate
to map. Also accepts dataframe with latitude in the first column and
longitude in the second column.}

\item{lng}{Numeric vector containing longitude coordinate for each
coordinate to map.}

\item{pat}{The API key used - this is usually aquired automatically through a helper
function}
}
\description{
Generate nearest point on the route network of a point using the CycleStreets.net
}
\section{Details}{

Retrieve coordinates of the node(s) on the network mapped from coordinates
passed to functions.
}
\examples{
\dontrun{
 nearest_cyclestreets(lat = 53, lng = 0.02, google_api = "api_key_here")
}
}

