% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google-functions.R
\name{nearest_google}
\alias{nearest_google}
\title{Generate nearest point on the route network of a point using the Google Maps API}
\usage{
nearest_google(lat, lng, google_api)
}
\arguments{
\item{lat}{Numeric vector containing latitude coordinate for each coordinate
to map. Also accepts dataframe with latitude in the first column and
longitude in the second column.}

\item{lng}{Numeric vector containing longitude coordinate for each
coordinate to map.}

\item{google_api}{String value containing the Google API key to use.}
}
\description{
Generate nearest point on the route network of a point using the Google Maps API
}
\section{Details}{

Retrieve coordinates of the node(s) on the network mapped from coordinates
passed to functions.
}

\examples{
\dontrun{
nearest_google(lat = 50.333, lng = 3.222, google_api = "api_key_here")
}
}
\seealso{
Other nodes: 
\code{\link{geo_code}()}
}
\concept{nodes}
