% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osrmfunctions.R
\name{nearest_osm}
\alias{nearest_osm}
\title{Generate nearest point on the route network of a point from OSRM locate service}
\usage{
nearest_osm(lat, lng, osrmurl = "http://router.project-osrm.org")
}
\arguments{
\item{lat}{Numeric vector containing latitude coordinate for each coordinate
to map. Also accepts dataframe with latitude in the first column and
longitude in the second column.}

\item{lng}{Numeric vector containing longitude coordinate for each
coordinate to map.}

\item{osrmurl}{Base URL of the OSRM service}
}
\description{
Generate nearest point on the route network of a point from OSRM locate service
}
\section{Details}{

Retrieve coordinates of the node(s) on the network mapped from coordinates
passed to functions.
}
\examples{
\dontrun{
 nearest_osm(
   lat = 50.3,
   lng = 13.2
 )
}
}

