% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/od-funs.R
\name{od_aggregate_from}
\alias{od_aggregate_from}
\title{Summary statistics of trips originating from zones in OD data}
\usage{
od_aggregate_from(flow, attrib = NULL, FUN = sum, ..., from_col = 1)
}
\arguments{
\item{flow}{A data frame representing the flow between two points
or zones. The first two columns of this data frame should correspond
to the first column of the data in the zones. Thus in \code{\link[=cents]{cents()}},
the first column is geo_code. This corresponds to the first two columns
of \code{\link[=flow]{flow()}}.}

\item{attrib}{A character vector corresponding to the variables in
\code{sl$} on which the function(s) will operate.}

\item{FUN}{A function to summarise OD data by}

\item{...}{Additional arguments passed to \code{FUN}}

\item{from_col}{The column that the OD dataset is grouped by (1 by default, the first column usually represents the origin)}
}
\description{
This function takes a data frame of OD data and
returns a data frame reporting summary statistics for each unique zone of origin.
}
\details{
It has some default settings: the default summary statistic is \code{sum()}.
If the third column is numeric, it returns a data frame with the total number of trips
originating from each zone, as illustrated in the examples below.
}
\examples{
od_aggregate_from(flow)
}
