% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/od_coords.R
\name{od_coords}
\alias{od_coords}
\title{General function to create a matrix representing origins and destinations}
\usage{
od_coords(from = NULL, to = NULL, l = NULL)
}
\arguments{
\item{from}{An object representing origins}

\item{to}{An object representing destinations}

\item{l}{If from and to are empty, this can be a spatial line object represent desire lines}
}
\description{
This function takes a wide range of input data types (spatial lines, points or text strings)
and returns a matrix of coordinates representing origin (fx, fy) and destination (tx, ty) points.
}
\examples{
od_coords(from = cents[1:3, ], to = cents[2:4, ]) # Spatial points
od_coords(cents_sf[1:3, ], cents_sf[2:4, ]) # sf points
od_coords("Bristol", "Gloucester")
# od_coords(c("Bristol", "Temple Meads"), c("Gloucester", "Bristol")) # not working
od_coords(l = flowlines)
# od_coords(l = flowlines_sf) # todo
}
