% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{od_data_lines}
\alias{od_data_lines}
\title{Example of desire line representations of origin-destination data from UK Census}
\format{
A data frame (tibble) object
}
\description{
Derived from \code{od_data_sample} showing movement between points represented in \code{cents_sf}
}
\examples{
od_data_lines
}
\seealso{
Other data: 
\code{\link{cents_sf}},
\code{\link{destinations_sf}},
\code{\link{flow_dests}},
\code{\link{flowlines_sf}},
\code{\link{flow}},
\code{\link{od_data_routes}},
\code{\link{od_data_sample}},
\code{\link{osm_net_example}},
\code{\link{read_table_builder}()},
\code{\link{route_network_sf}},
\code{\link{route_network_small}},
\code{\link{routes_fast_sf}},
\code{\link{routes_slow_sf}},
\code{\link{zones_sf}}
}
\concept{data}
\keyword{datasets}
