% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{od_data_routes}
\alias{od_data_routes}
\title{Example segment-level route data}
\format{
A data frame (tibble) object
}
\description{
See \code{data-raw/generate-data.Rmd} for details on how this was created.
The dataset shows routes between origins and destinations represented in
\code{od_data_lines}
}
\examples{
od_data_routes
}
\seealso{
Other data: 
\code{\link{cents_sf}},
\code{\link{destinations_sf}},
\code{\link{flow_dests}},
\code{\link{flowlines_sf}},
\code{\link{flow}},
\code{\link{od_data_lines}},
\code{\link{od_data_sample}},
\code{\link{osm_net_example}},
\code{\link{read_table_builder}()},
\code{\link{route_network_sf}},
\code{\link{routes_fast_sf}},
\code{\link{routes_slow_sf}},
\code{\link{zones_sf}}
}
\concept{data}
\keyword{datasets}
