% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{od_data_sample}
\alias{od_data_sample}
\title{Example of origin-destination data from UK Census}
\format{A data frame (tibble) object}
\description{
See \code{data-raw/generate-data.Rmd} for details on how this was created.
}
\examples{
od_data_sample
}
\keyword{datasets}
