% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oneway.R
\name{od_oneway}
\alias{od_oneway}
\title{Aggregate od pairs they become non-directional}
\usage{
od_oneway(x, attrib, id1 = names(x)[1], id2 = names(x)[2],
  stplanr.key = NULL)
}
\arguments{
\item{x}{A data frame or SpatialLinesDataFrame, representing an OD matrix}

\item{attrib}{A vector of column numbers or names
for deciding which attribute(s) of class numeric to
aggregate}

\item{id1}{Optional (it is assumed to be the first column)
text string referring to the name of the variable containing
the unique id of the origin}

\item{id2}{Optional (it is assumed to be the second column)
text string referring to the name of the variable containing
the unique id of the destination}

\item{stplanr.key}{Optional key of unique OD pairs regardless of the order,
e.g., as generated by \code{\link[=od_id_max_min]{od_id_max_min()}} or \code{\link[=od_id_szudzik]{od_id_szudzik()}}}
}
\value{
\code{oneway} outputs a data frame (or \code{sf} data frame) with rows containing
results for the user-selected attribute values that have been aggregated.
}
\description{
For example, sum total travel in both directions.
}
\details{
Flow data often contains movement in two directions: from point A to point B
and then from B to A. This can be problematic for transport planning, because
the magnitude of flow along a route can be masked by flows the other direction.
If only the largest flow in either direction is captured in an analysis, for
example, the true extent of travel will by heavily under-estimated for
OD pairs which have similar amounts of travel in both directions.
Flows in both direction are often represented by overlapping lines with
identical geometries (see \code{\link[=flowlines]{flowlines()}}) which can be confusing
for users and are difficult to plot.
}
\examples{
flow_oneway <- od_oneway(flow, attrib = 3)
nrow(flow_oneway) < nrow(flow) # result has fewer rows
sum(flow$All) == sum(flow_oneway$All) # but the same total flow
# using names instead of index for attribute
od_oneway(flow, attrib = "All")
# using many attributes to aggregate
attrib <- which(vapply(flow, is.numeric, TRUE))
flow_oneway <- od_oneway(flow, attrib = attrib)
colSums(flow_oneway[attrib]) == colSums(flow[attrib]) # test if the colSums are equal
# Demonstrate the results from oneway and onewaygeo are identical
flow_oneway_geo <- onewaygeo(flowlines, attrib = attrib)
plot(flow_oneway$All, flow_oneway_geo$All)
flow_oneway_sf <- od_oneway(flowlines_sf, attrib)
sum(flow_oneway$All) == sum(flow_oneway_sf$All) # but the same total flow
summary(flow_oneway$All \%in\% flow_oneway_sf$All)
plot(flow_oneway_geo, lwd = flow_oneway_geo$All / mean(flow_oneway_geo$All))
plot(flow_oneway_sf$geometry, lwd = flow_oneway_sf$All / mean(flow_oneway_sf$All))
fr <- flowlines_sf[[1]]
to <- flowlines_sf[[2]]
flow_max_min <- od_oneway(flowlines_sf, attrib, stplanr.key = od_id_character(fr, to))
cor(flow_max_min$All, flow_oneway_sf$All)
# benchmark performance
bench::mark(check = FALSE,
  onewayid(flowlines_sf, attrib),
  od_oneway(flowlines_sf, attrib),
  od_oneway(flowlines_sf, attrib, stplanr.key = od_id_character(fr, to))
)
}
\seealso{
Other lines: \code{\link{angle_diff}},
  \code{\link{geo_toptail}}, \code{\link{is_linepoint}},
  \code{\link{line2df}}, \code{\link{line_bearing}},
  \code{\link{line_match}}, \code{\link{line_midpoint}},
  \code{\link{line_sample}}, \code{\link{line_segment}},
  \code{\link{line_via}}, \code{\link{mats2line}},
  \code{\link{n_sample_length}}, \code{\link{n_vertices}},
  \code{\link{onewaygeo}}, \code{\link{onewayid}},
  \code{\link{points2line}}, \code{\link{toptail_buff}},
  \code{\link{toptailgs}},
  \code{\link{update_line_geometry}}
}
\concept{lines}
