% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiate.R
\name{od_radiation}
\alias{od_radiation}
\title{Function which estimates the flow between points/zones using the 'radiation model'}
\usage{
od_radiation(p, pop_var = "population")
}
\arguments{
\item{p}{A SpatialPoints dataframe, the first column of which contains it's unique ID}

\item{pop_var}{A character string representing the variable that corresponds
to the population of the zone/point}
}
\description{
Function which estimates the flow between points/zones using the 'radiation model'
}
\examples{
data(cents)
plot(cents)
class(cents)
# Create test population to model flows
set.seed(2050)
cents$population <- runif(n = nrow(cents), min = 100, max = 1000)
flowlines_radiation <- od_radiation(cents, pop_var = "population")
plot(flowlines_radiation, lwd = flowlines_radiation$flow * 10)
}

