% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oneway.R
\name{od_szudzik_pairing}
\alias{od_szudzik_pairing}
\title{Combines two ID values to create a single ID number}
\usage{
od_szudzik_pairing(val1, val2, ordermatters = FALSE)
}
\arguments{
\item{val1}{a vector of numeric, character, or factor values}

\item{val2}{a vector of numeric, character, or factor values}

\item{ordermatters}{logical, does the order of values matter to pairing, default = FALSE}
}
\description{
Combines two ID values to create a single ID number
}
\details{
In OD data it is common to have many flows from "A to B" and "B to A".
It can be useful to group these an have a single ID that represents pairs of IDs
with or without directionality.

This function implements the Szudzik pairing function, on two vectors of equal
length. It returns a vector of ID numbers.

This function superseeds od_id_order as it is faster on large datasets
}
\examples{

szudzik_pairing(od_data_sample[[1]], od_data_sample[[2]])
od_id_order(od_data_sample)
n = 1000
ids <- as.character(runif(n, 1e4, 1e7 - 1))
x <- data.frame(id1 = rep(ids, times = n),
                id2 = rep(ids, each = n),
                val = 1,
                stringsAsFactors = FALSE)
system.time(od_id_order(x))
system.time(szudzik_pairing(x$id1, x$id2))
}
