% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overline_sf.R
\name{overline_sf}
\alias{overline_sf}
\title{Convert series of overlapping lines into a route network}
\usage{
overline_sf(sl, attrib, fun = sum)
}
\arguments{
\item{sl}{An \code{sf} \code{LINESTRING} object with overlapping elements}

\item{attrib}{A character vector corresponding to the variables in
\code{sl$} on which the function(s) will operate.}

\item{fun}{The function(s) used to aggregate the grouped values (default: sum).
If length of \code{fun} is smaller than \code{attrib} then the functions are
repeated for subsequent attributes.}
}
\description{
This function takes overlapping \code{LINESTRING}s stored in an
\code{sf} object and returns a route network composed of non-overlapping
geometries and aggregated values.
}
\examples{
\dontrun{
routes_fast_sf$value = 1
sl <- routes_fast_sf[2:5, ]
rnet = overline_sf(sl = sl, attrib = c("value", "length"))
plot(rnet, lwd = rnet$value)
# rnet_sf <- overline(routes_fast_sf, attrib = "length", buff_dist = 10)
# plot(rnet_sf$geometry, lwd = rnet_sf$length / mean(rnet_sf$length))
}
}
