% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overline_sf.R
\name{overline_sf2}
\alias{overline_sf2}
\title{Combine 2 \code{sf} \code{LINESTRING} objects into non-overlapping segments and
find the aggregate value of selected columns defined by \code{attrib} along shared segments.}
\usage{
overline_sf2(sl, sl2, attrib = "value", fun = sum,
  return_linestring = FALSE)
}
\arguments{
\item{sl}{A SpatialLinesDataFrame with overlapping elements}

\item{sl2}{A second \code{LINESTRING} object to aggregate}

\item{attrib}{A character vector corresponding to the variables in
\code{sl$} on which the function(s) will operate.}

\item{fun}{The function(s) used to aggregate the grouped values (default: sum).
If length of \code{fun} is smaller than \code{attrib} then the functions are
repeated for subsequent attributes.}

\item{return_linestring}{Can the result include \code{MULTILINESTRING} geometries?
Default: \code{FALSE}}
}
\description{
Combine 2 \code{sf} \code{LINESTRING} objects into non-overlapping segments and
find the aggregate value of selected columns defined by \code{attrib} along shared segments.
}
\examples{
routes_fast_sf$value = 1
sl = routes_fast_sf[2, ]
sl2 = routes_fast_sf[3, ]
rnet = overline_sf2(sl, sl2, attrib = "value")
unique(sf::st_geometry_type(rnet))
plot(rnet, lwd = rnet$value)
rnet2 = overline_sf2(sl, sl2, attrib = c("value", "length"))
plot(rnet2)
rnet = overline_sf(routes_fast_sf[2:6, "value"], "value")
plot(rnet$geometry, lwd = rnet$value)
plot(routes_fast_sf[7, ], add = TRUE, col = "red")
sel_o = overlaps(rnet, routes_fast_sf[7, ])
rnet_sub = rnet[sel_o, ]
plot(rnet_sub$geometry, lwd = rnet_sub$value)
plot(routes_fast_sf[7, ], add = TRUE, col = "red")
rnet_new = overline_sf2(rnet_sub[1, ], routes_fast_sf[7, ])
plot(rnet_new$geometry, lwd = 8)
plot(rnet_new[1, ], add = TRUE, col = "red")
plot(rnet_new[2, ], add = TRUE, col = "red")
plot(rnet_new[3, ], add = TRUE, col = "red")
plot(rnet_new[4, ], add = TRUE, col = "red")
}
