% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overline.R
\name{overline_Spatial}
\alias{overline_Spatial}
\title{Spatial aggregation on sp data}
\usage{
overline_Spatial(sl, attrib, fun = sum, na.zero = FALSE, buff_dist = 0)
}
\arguments{
\item{sl}{SpatialLinesDataFrame with overlapping Lines to split by
number of overlapping features.}

\item{attrib}{character, column names in sl to be aggregated}

\item{fun}{Named list of functions to summaries the attributes by? \code{sum} is the default.
\code{list(sum = sum, average = mean)} will summarise all \code{attrib}utes by sum and mean.}

\item{na.zero}{Sets whether aggregated values with a value of zero are
removed.}

\item{buff_dist}{A number specifying the distance in meters of the buffer to be used to crop lines before running the operation.
If the distance is zero (the default) touching but non-overlapping lines may be aggregated.}
}
\description{
This function, largely superceded by sf implementations, still works
but is not particularly fast.
}
