% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialLinesNetwork.R
\docType{methods}
\name{plot,sfNetwork,ANY-method}
\alias{plot,sfNetwork,ANY-method}
\title{Plot a sfNetwork}
\usage{
\S4method{plot}{sfNetwork,ANY}(x, component = "sl", ...)
}
\arguments{
\item{x}{The sfNetwork to plot}

\item{component}{The component of the network to plot. Valid values are "sl"
for the geographic (sf) representation or "graph" for the graph
representation.}

\item{...}{Arguments to pass to relevant plot function.}
}
\description{
Plot a sfNetwork
}
\examples{
data(routes_fast)
rnet <- overline(routes_fast, attrib = "length")
SLN <- SpatialLinesNetwork(rnet)
plot(SLN)
}
