% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialLinesNetwork.R
\docType{methods}
\name{plot,SpatialLinesNetwork,ANY-method}
\alias{plot,SpatialLinesNetwork,ANY-method}
\title{Plot a SpatialLinesNetwork}
\usage{
\S4method{plot}{SpatialLinesNetwork,ANY}(x, component = "sl", ...)
}
\arguments{
\item{x}{The SpatialLinesNetwork to plot}

\item{component}{The component of the network to plot. Valid values are "sl"
for the geographic (SpatialLines) representation or "graph" for the graph
representation.}

\item{...}{Arguments to pass to relevant plot function.}
}
\description{
Plot a SpatialLinesNetwork
}

