% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialLinesNetwork.R
\name{plot.SpatialLinesNetwork}
\alias{plot.SpatialLinesNetwork}
\title{Plot a SpatialLinesNetwork}
\usage{
plot.SpatialLinesNetwork(sln, component = "sl", ...)
}
\arguments{
\item{sln}{The SpatialLinesNetwork to plot}

\item{component}{The component of the network to plot. Valid values are "sl"
for the geographic (SpatialLines) representation or "graph" for the graph
representation.}
}
\description{
Plot a SpatialLinesNetwork
}

