% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/loadABS.R
\name{readTableBuilder}
\alias{readTableBuilder}
\title{Import and format Australian Bureau of Statistics (ABS) TableBuilder files}
\usage{
readTableBuilder(dataset, filetype = "csv", sheet = 1, removeTotal = TRUE)
}
\arguments{
\item{dataset}{Either a dataframe containing the original data from
TableBuilder or a character string containing the path of the
unzipped TableBuilder file.}

\item{filetype}{A character string containing the filetype. Valid values
are 'csv' and 'xlsx' (default = 'csv'). Not used if dataset is a dataframe.}

\item{sheet}{An integer value containing the index of the sheet in the
xlsx file (default = 1).}

\item{removeTotal}{A boolean value. If TRUE removes the rows and columns
with totals (default = TRUE).}
}
\description{
Import and format Australian Bureau of Statistics (ABS) TableBuilder files
}
\section{Details}{

The Australian Bureau of Statistics (ABS) provides customised tables for
census and other datasets in a format that is difficult to use in R
because it contains rows with additional information.
This function imports the original (unzipped) TableBuilder files in .csv
or .xlsx format before creating an R dataframe with the data.
}
\examples{
\dontrun{
readTableBuilder('employmentSLA.csv')
readTableBuilder('employmentSLA.xlsx',filetype='xlsx',sheet=1,removeTotal=TRUE)
employment <- read.csv('employmentSLA.csv')
readTableBuilder(employment)
}
}

