% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnet-clean.R
\name{rnet_clean_vertices}
\alias{rnet_clean_vertices}
\title{Cleanr a sf object with LINESTRING geometry such that}
\usage{
rnet_clean_vertices(rnet)
}
\arguments{
\item{rnet}{a sf LINESTRING object that ...}
}
\value{
sf object
}
\description{
Cleanr a sf object with LINESTRING geometry such that
}
\examples{
library(sf)
par(mar = rep(0, 4))

# Check for roundabout
plot(rnet_roundabout$geometry, lwd = 2, col = rainbow(nrow(rnet_roundabout)))

rnet_roundabout_clean <- rnet_clean_vertices(rnet_roundabout)
plot(rnet_roundabout_clean$geometry, lwd = 2, col = rainbow(nrow(rnet_roundabout_clean)))
# Check for overpasses
plot(rnet_overpass$geometry, lwd = 2, col = rainbow(nrow(rnet_overpass)))

rnet_overpass_clean <- rnet_clean_vertices(rnet_overpass)
plot(rnet_overpass_clean$geometry, lwd = 2, col = rainbow(nrow(rnet_overpass_clean)))
#' \donttest{
mapview::mapview(rnet_overpass_clean)
}
# Check for intersection with no node
plot(rnet_cycleway_intersection$geometry, lwd = 2, col = rainbow(nrow(rnet_cycleway_intersection)))

rnet_cycleway_intersection_clean <- rnet_clean_vertices(rnet_cycleway_intersection)
plot(rnet_cycleway_intersection_clean$geometry, lwd = 2, col = rainbow(nrow(rnet_cycleway_intersection_clean)))

# Bigger example
\donttest{
library(geofabric)
iow <- geofabric::get_geofabric("isle wight")

key_roads_text = "primary|secondary|tertiary|cycleway|trunk|motorway"
iow_small <- iow[grepl(pattern = key_roads_text, x = iow$highway), ]
system.time(iow_clean <- rnet_clean_vertices(iow_small))

plot(iow_small$geometry)
plot(iow_clean$geometry)
}
}
