% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnet_group.R
\name{rnet_connected}
\alias{rnet_connected}
\title{Keep only segments connected to the largest group in a network}
\usage{
rnet_connected(rnet)
}
\arguments{
\item{rnet}{An sf object representing a road network}
}
\value{
An sf object representing the largest group in the network
}
\description{
This function takes an sf object representing a road network and
returns only the parts of the network that are in the largest group.
}
\examples{
rnet <- rnet_breakup_vertices(stplanr::osm_net_example)
rnet_largest_group <- rnet_connected(rnet)
plot(rnet$geometry)
plot(rnet_largest_group$geometry)
}
