% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rnet_cycleway_intersection}
\alias{rnet_cycleway_intersection}
\title{Example of cycleway intersection data showing problems for SpatialLinesNetwork objects}
\format{A sf object}
\description{
See \code{data-raw/rnet_cycleway_intersection} for details on how this was created.
}
\examples{
rnet_cycleway_intersection
}
\keyword{datasets}
