% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node-funs.R
\name{rnet_get_nodes}
\alias{rnet_get_nodes}
\title{Extract nodes from route network}
\usage{
rnet_get_nodes(rnet, p = NULL)
}
\arguments{
\item{rnet}{A route network of the type generated by \code{overline()}}

\item{p}{A point represented by an \code{sf} object the will split the \code{route}}
}
\description{
Extract nodes from route network
}
\examples{
rnet_get_nodes(route_network_sf)
}
