% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnet_junctions.R
\name{rnet_junction}
\alias{rnet_junction}
\title{Get points where 3 or more linestrings meet (junctions on road networks)}
\usage{
rnet_junction(rnet)
}
\arguments{
\item{rnet}{An sf object with LINESTRING geometry representing a route
network.}
}
\description{
Get points where 3 or more linestrings meet (junctions on road networks)
}
\examples{
has_sfheaders <- requireNamespace("sfheaders", quietly = TRUE)
if(has_sfheaders) {
rnet <- rnet_roundabout
j <- rnet_junction(rnet)
plot(rnet$geometry)
plot(j, add = TRUE)
}
}
