% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rnet_overpass}
\alias{rnet_overpass}
\title{Example of overpass data showing problems for SpatialLinesNetwork objects}
\format{A sf object}
\description{
See \code{data-raw/rnet_overpass.R} for details on how this was created.
}
\examples{
rnet_overpass
}
\keyword{datasets}
