% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rnet_roundabout}
\alias{rnet_roundabout}
\title{Example of roundabout data showing problems for SpatialLinesNetwork objects}
\format{A sf object}
\description{
See \code{data-raw/rnet_roundabout.R} for details on how this was created.
}
\examples{
rnet_roundabout
}
\keyword{datasets}
