% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnet_join.R
\name{rnet_split_lines}
\alias{rnet_split_lines}
\title{Split lines in a route network based points}
\usage{
rnet_split_lines(rnet_x, geo_y, dist = 1)
}
\arguments{
\item{rnet_x}{The route network to be broken into smaller pieces}

\item{geo_y}{The geographic object used to break up the route network}

\item{dist}{The width of the buffer used when breaking up the route network.
For imprecise data it may be worth increasing this above 1 m, the default.}
}
\description{
If the object passed to the second argument has LINSTRING geometries
the start and end points of linestrings are used.
}
