% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnet_join.R
\name{rnet_subset}
\alias{rnet_subset}
\title{Subset one route network based on overlaps with another}
\usage{
rnet_subset(rnet_x, rnet_y, dist = 1, crop = TRUE, min_x = 3)
}
\arguments{
\item{rnet_x}{The route network to be subset}

\item{rnet_y}{The subsetting route network}

\item{dist}{The buffer width around y in meters. 1 m by default.}

\item{crop}{Crop \code{rnet_x}? \code{TRUE} is the default}

\item{min_x}{Segments shorter than this multiple of dist
\emph{and} which were longer
before the cropping process will be removed. 3 by default.}
}
\description{
Subset one route network based on overlaps with another
}
