% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route.R
\name{route_dodgr}
\alias{route_dodgr}
\title{Route on local data using the dodgr package}
\usage{
route_dodgr(from = NULL, to = NULL, l = NULL, rnet = NULL)
}
\arguments{
\item{from}{An object representing origins
(if lines are provided as the first argument, from is assigned to \code{l})}

\item{to}{An object representing destinations}

\item{l}{Only needed if from and to are empty, in which case this
should be a spatial object representing desire lines}

\item{rnet}{sf object representing the route network}
}
\description{
Route on local data using the dodgr package
}
\examples{
from <- sf::st_coordinates(cents_sf[1, ])
to <- sf::st_coordinates(cents_sf[2, ])
r <- route_dodgr(from, to)
}
