% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/routes.R
\name{route_graphhopper}
\alias{route_graphhopper}
\title{Plan a route with the graphhopper routing engine}
\usage{
route_graphhopper(from, to, vehicle = "bike")
}
\arguments{
\item{from}{Text string or coordinates (a numeric vector of
 \code{length = 2} representing latitude and longitude) representing a point
 on Earth.}

\item{to}{Text string or coordinates (a numeric vector of
 \code{length = 2} representing latitude and longitude) representing a point
 on Earth. This represents the destination of the trip.}

\item{vehicle}{A text string representing the vehicle. Can be bike, bike2, car
or foot.}
}
\description{
Provides an R interface to the graphhopper routing engine,
an open source route planning service.
To use this function you will need to obtain an API key from
\url{https://graphhopper.com/#directions-api}.
It is assumed that you have set your api key as a system environment
for security reasons (so you avoid typing the API key in your code).
Do this with \code{GRAPHHOPPER}, e.g. by typing \code{Sys.setenv(GRAPHHOPPER = 'eccbf612-214e-437d-8b73-06bdf9e68731')}. (Note: key not real, use your own key.)
}
\details{
The function returns a SpatialLinesDataFrame object.
See \url{https://github.com/graphhopper} for more information.

To test graphopper is working for you, try something like this, but with
your own API key:

\code{obj <- jsonlite::fromJSON(url)}

Where \code{url} is an example api request from \url{https://github.com/graphhopper/directions-api/blob/master/docs-routing.md}.
}
\examples{
\dontrun{
r <- route_graphhopper("Leeds", "Dublin", vehicle = "bike")

library(leaflet)

leaflet() \%>\% addTiles() \%>\% addPolylines(data = r)
}
}
\seealso{
route_cyclestreet
}

