% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{route_network}
\alias{route_network}
\alias{route_network_sf}
\title{spatial lines dataset representing a route network}
\format{A spatial lines dataset 80 rows and 1 column}
\usage{
data(route_network)
}
\description{
spatial lines dataset representing a route network
}
\examples{
\dontrun{
# Generate route network
route_network = overline(routes_fast, "All", fun = sum)
route_network_sf <- sf::st_as_sf(route_network)
}
}
\keyword{datasets}
