% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{route_network_sf}
\alias{route_network_sf}
\title{Spatial lines dataset representing a route network}
\format{
A spatial lines dataset 80 rows and 1 column
}
\description{
The flow of commuters using different segments of the road network represented in the
\code{\link[=flowlines_sf]{flowlines_sf()}} and \code{\link[=routes_fast_sf]{routes_fast_sf()}} datasets
}
\seealso{
Other data: 
\code{\link{cents_sf}},
\code{\link{destinations_sf}},
\code{\link{flow}},
\code{\link{flow_dests}},
\code{\link{flowlines_sf}},
\code{\link{od_data_lines}},
\code{\link{od_data_routes}},
\code{\link{od_data_sample}},
\code{\link{osm_net_example}},
\code{\link{read_table_builder}()},
\code{\link{route_network_small}},
\code{\link{routes_fast_sf}},
\code{\link{routes_slow_sf}},
\code{\link{zones_sf}}
}
\concept{data}
\keyword{datasets}
