% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{route_network_small}
\alias{route_network_small}
\title{Spatial lines dataset representing a small route network}
\format{
A spatial lines dataset with one column: flow
}
\description{
The flow between randomly selected vertices on the \code{osm_net_example}.
See \code{data-raw/route_network_small.R} for details.
}
\seealso{
Other data: 
\code{\link{cents_sf}},
\code{\link{destinations_sf}},
\code{\link{flow}},
\code{\link{flow_dests}},
\code{\link{flowlines_sf}},
\code{\link{od_data_lines}},
\code{\link{od_data_routes}},
\code{\link{od_data_sample}},
\code{\link{osm_net_example}},
\code{\link{read_table_builder}()},
\code{\link{route_network_sf}},
\code{\link{routes_fast_sf}},
\code{\link{routes_slow_sf}},
\code{\link{zones_sf}}
}
\concept{data}
\keyword{datasets}
