% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slope.R
\name{route_slope_vector}
\alias{route_slope_vector}
\title{Calculate the gradient of line segments from distance and elevation vectors}
\usage{
route_slope_vector(x, e)
}
\arguments{
\item{x}{Vector of locations}

\item{e}{Elevations in same units as x (assumed to be metres)}
}
\description{
Calculate the gradient of line segments from distance and elevation vectors
}
\examples{
x <- c(0, 2, 3, 4, 5, 9)
e <- c(1, 2, 2, 4, 3, 1) / 10
route_slope_vector(x, e)
}
\seealso{
Other route_funs: 
\code{\link{route_average_gradient}()},
\code{\link{route_rolling_average}()},
\code{\link{route_rolling_diff}()},
\code{\link{route_rolling_gradient}()},
\code{\link{route_sequential_dist}()},
\code{\link{route_slope_matrix}()}
}
\concept{route_funs}
