% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route-transport-api.R
\name{route_transportapi}
\alias{route_transportapi}
\title{Plan a single route with TransportAPI.com}
\usage{
route_transportapi(from, to, silent = FALSE,
  base_url = "http://fcc.transportapi.com/v3/uk")
}
\arguments{
\item{from}{Text string or coordinates (a numeric vector of
\code{length = 2} representing latitude and longitude) representing a point
on Earth.}

\item{to}{Text string or coordinates (a numeric vector of
\code{length = 2} representing latitude and longitude) representing a point
on Earth. This represents the destination of the trip.}

\item{silent}{Logical (default is FALSE). TRUE hides request sent.}
}
\description{
Provides an R interface to the TransportAPI.com public transport API.
See \url{https://developer.transportapi.com/documentation}for more information.
}
\details{
This function uses the online routing service
TransportAPI.com to find public routes
between origins and destinations. It does not require
any key to access the API.

Note that if \code{from} and \code{to} are supplied as
character strings (instead of lon/lat pairs), Google's
geo-coding services are used via \code{geo_code}.
}
\examples{

\dontrun{
route_transportapi("leeds uk", "bradford")
}

}
\seealso{
line2route
}
