% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{routes_fast}
\alias{routes_fast}
\title{SpatialLinesDataFrame of commuter flows on the travel network}
\format{A SpatialLinesDataFrame 49 rows and 15 columns}
\usage{
data(routes_fast)
}
\description{
SpatialLinesDataFrame of commuter flows on the travel network
}
\seealso{
Other res: \code{\link{flowlines}}; \code{\link{flow}};
  \code{\link{routes_slow}}
}
\keyword{datasets}

