% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{routes_fast_sf}
\alias{routes_fast_sf}
\title{Spatial lines dataset of commuter flows on the travel network}
\format{
A spatial lines dataset with 49 rows and 15 columns
}
\usage{
routes_fast_sf
}
\description{
Simulated travel route allocated to the transport network
representing the 'fastest' between \code{cents_sf}
objects.
}
\seealso{
Other example data: 
\code{\link{destinations_sf}},
\code{\link{flow_dests}},
\code{\link{flowlines_sf}},
\code{\link{flow}},
\code{\link{route_network_sf}},
\code{\link{routes_slow_sf}}
}
\concept{example data}
\keyword{datasets}
