% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{routes_quiet}
\alias{routes_quiet}
\title{SpatialLinesDataFrame of commuter flows on the travel network}
\format{A SpatialLinesDataFrame 49 rows and 15 columns}
\usage{
data(routes_slow)
}
\description{
Simulated travel route allocated to the transport network
representing the 'quietest' between \code{\link{cents}()}
objects
with \code{\link{gFlow2line}} (see \code{\link{flow}}).
}
\seealso{
\code{\link{gFlow2line}}  \code{\link{cents}}
 \code{\link{flow}}  \code{\link{routes_slow}}
  \code{\link{routes_fast}}
}
\keyword{datasets}

