% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{routes_slow}
\alias{routes_slow}
\title{SpatialLinesDataFrame of commuter flows on the travel network}
\format{A SpatialLinesDataFrame 49 rows and 15 columns}
\usage{
data(routes_slow)
}
\description{
Simulated travel route allocated to the transport network
representing the 'quietest' between \code{\link{cents}()}
objects
with \code{\link{gFlow2line}} (see \code{\link{flow}}).
}
\seealso{
Other example.flow.data: \code{\link{cents}};
  \code{\link{flowlines}}; \code{\link{flow}};
  \code{\link{routes_fast}}
}
\keyword{datasets}

