% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{routes_slow_sf}
\alias{routes_slow_sf}
\title{Spatial lines dataset of commuter flows on the travel network}
\format{
A spatial lines dataset 49 rows and 15 columns
}
\description{
Simulated travel route allocated to the transport network
representing the 'quietest' between \code{cents_sf}.
}
\seealso{
Other example data: 
\code{\link{destinations_sf}},
\code{\link{flow_dests}},
\code{\link{flowlines_sf}},
\code{\link{flow}},
\code{\link{route_network_sf}},
\code{\link{routes_fast_sf}}
}
\concept{example data}
\keyword{datasets}
