% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialLinesNetwork.R
\docType{class}
\name{sfNetwork-class}
\alias{sfNetwork-class}
\title{An S4 class representing a (typically) transport network}
\description{
This class uses a combination of a sf layer and an igraph
object to represent transport networks that can be used for routing and
other network analyses.
}
\section{Slots}{

\describe{
\item{\code{sl}}{A sf line layer with the geometry and other attributes
for each link the in network.}

\item{\code{g}}{The graph network corresponding to \code{sl}.}

\item{\code{nb}}{A list containing vectors of the nodes connected to each node
in the network.}

\item{\code{weightfield}}{A character vector containing the variable (column) name
from the SpatialLinesDataFrame to be used for weighting the network.}
}}

