% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialLinesNetwork.R
\name{sln2points}
\alias{sln2points}
\title{Generate spatial points representing nodes on a SpatialLinesNetwork
or sfNetwork.}
\usage{
sln2points(sln)
}
\arguments{
\item{sln}{The SpatialLinesNetwork to use.}
}
\description{
Generate spatial points representing nodes on a SpatialLinesNetwork
or sfNetwork.
}
\examples{
data(routes_fast)
rnet <- overline(routes_fast, attrib = "length")
SLN <- SpatialLinesNetwork(rnet)
(sln_nodes <- sln2points(SLN))
plot(SLN)
plot(sln_nodes, add = TRUE)
}
