% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialLinesNetwork.R
\name{sln_clean_graph}
\alias{sln_clean_graph}
\title{Clean spatial network - return an sln with a single connected graph}
\usage{
sln_clean_graph(sln)
}
\arguments{
\item{sln}{A spatial lines (\code{sfNetwork}) object created by \code{SpatialLinesNetwork}}
}
\value{
An sfNetwork object
}
\description{
See https://github.com/ropensci/stplanr/issues/344
}
