% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depreciated.R
\name{stplanr-deprecated}
\alias{stplanr-deprecated}
\title{Deprecated functions in stplanr}
\description{
These functions are depreciated and will be removed:
}
\details{
\itemize{
\item \code{\link[=dl_stats19]{dl_stats19()}}: see the \href{https://github.com/ropensci/stats19}{stats19} package
}
}
