% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialLinesNetwork.R
\name{summary,sfNetwork-method}
\alias{summary,sfNetwork-method}
\title{Print a summary of a sfNetwork}
\usage{
\S4method{summary}{sfNetwork}(object, ...)
}
\arguments{
\item{object}{The sfNetwork}

\item{...}{Arguments to pass to relevant summary function.}
}
\description{
Print a summary of a sfNetwork
}
\examples{
data(routes_fast)
rnet <- overline(routes_fast, attrib = "length")
sln <- SpatialLinesNetwork(rnet)
summary(sln)
}
