% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialLinesNetwork.R
\docType{methods}
\name{summary,SpatialLinesNetwork-method}
\alias{summary,SpatialLinesNetwork-method}
\title{Print a summary of a SpatialLinesNetwork}
\usage{
\S4method{summary}{SpatialLinesNetwork}(object, ...)
}
\arguments{
\item{object}{The SpatialLinesNetwork}

\item{...}{Arguments to pass to relevant plot function.}
}
\description{
Print a summary of a SpatialLinesNetwork
}

