% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialLinesNetwork.R
\name{summary.SpatialLinesNetwork}
\alias{summary.SpatialLinesNetwork}
\title{Print a summary of a SpatialLinesNetwork}
\usage{
summary.SpatialLinesNetwork(sln, ...)
}
\arguments{
\item{sln}{The SpatialLinesNetwork}
}
\description{
Print a summary of a SpatialLinesNetwork
}

