% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osrmfunctions.R
\name{table2matrix}
\alias{table2matrix}
\title{Return SpatialPointsDataFrame with nearest street from OSRM nearest service}
\usage{
table2matrix(lat, lng = NA, osrmurl = "http://router.project-osrm.org")
}
\arguments{
\item{lat}{Numeric vector containing latitude coordinate for each coordinate
to calculate travel times. Also accepts dataframe with latitude in the first
column and longitude in the second column.}

\item{lng}{Numeric vector containing longitude coordinate for each
coordinate to calculate travel times.}

\item{osrmurl}{Base URL of the OSRM service}
}
\description{
Return SpatialPointsDataFrame with nearest street from OSRM nearest service
}
\section{Details}{

Retrieve coordinates and name of the node(s) on the network mapped from
coordinates passed to functions.
}
\examples{
\dontrun{
 table2matrix(seq(from=50,to=52,by=0.1),seq(from=12,to=14,by=0.1))
}

}

