% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overline_sf.R
\name{to_linestring}
\alias{to_linestring}
\title{Identify overlapping lines}
\usage{
to_linestring(sl)
}
\arguments{
\item{sl}{A SpatialLinesDataFrame with overlapping elements}
}
\description{
Convert geometries that may be a mixture of \code{MULTILINESTRING}s
and \code{LINESTRING}s into non-overlapping \code{LINESTRING}s.
}
\examples{
l1 = sf::st_linestring(matrix(c(1, 2, 4, 1, 1, 2), ncol = 2))
l2 = sf::st_linestring(matrix(c(1, 2, 4, 1, 1, 0), ncol = 2))
ml = sf::st_multilinestring(list(l1, l2))
sl = sf::st_sf(sf::st_sfc(l1, l2, ml))
to_linestring(sl)
}
