% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toptail.R
\name{toptail}
\alias{toptail}
\title{Clip the first and last n metres of SpatialLines}
\usage{
toptail(l, toptail_dist)
}
\arguments{
\item{l}{A SpatialLines object}

\item{toptail_dies}{The distance (in metres) to top and tail the line by}
}
\description{
Takes lines and removes the start and end point, to a distance determined
by the user.
}
\examples{
data("routes_fast")
sp::proj4string(routes_fast) <- CRS("+init=epsg:4326")
r_toptail <- toptail(routes_fast, toptail_dist = 300)
plot(routes_fast, lwd = 3)
plot(r_toptail, col = "red", add = TRUE)
plot(cents, add = TRUE)
# Note the behaviour when the buffer size removes lines
r_toptail <- toptail(routes_fast, toptail_dist = 1000)
plot(routes_fast, lwd = 3)
plot(r_toptail, col = "red", add = TRUE)
}

