% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{wb}
\alias{wb}
\title{list of data labels associated with UK 'Stats19' road casualty data, used to clean
the raw data}
\format{A list of 47 dataframes}
\usage{
data(wb)
}
\description{
These labels originate from this project: \url{https://github.com/Robinlovelace/bikeR}
}
\keyword{datasets}

