% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zones_sf}
\alias{zones_sf}
\title{Spatial polygons of home locations for flow analysis.}
\description{
These correspond to the \code{cents_sf} data.
}
\details{
\itemize{
\item geo_code. the official code of the zone
}
}
\examples{
library(sf)
zones_sf
plot(zones_sf)
}
\keyword{datasets}
