% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_sf_fun.R
\name{as_sf_fun}
\alias{as_sf_fun}
\alias{as_sp_fun}
\title{Convert functions support sf/sp}
\usage{
as_sf_fun(input, FUN, ...)
}
\arguments{
\item{input}{Input object - an sf or sp object}

\item{FUN}{A function that works on sp/sf data}

\item{...}{Arguments passed to \code{FUN}}
}
\description{
Convert functions support sf/sp
}
