% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linefuns.R
\name{bind_sf}
\alias{bind_sf}
\title{Rapid row-binding of sf objects}
\usage{
bind_sf(x)
}
\arguments{
\item{x}{List of sf objects to combine}
}
\value{
An sf data frame
}
\description{
Rapid row-binding of sf objects
}
\seealso{
Other geo: 
\code{\link{bbox_scale}()},
\code{\link{geo_bb}()},
\code{\link{geo_bb_matrix}()},
\code{\link{geo_buffer}()},
\code{\link{geo_length}()},
\code{\link{geo_projected}()},
\code{\link{geo_select_aeq}()},
\code{\link{quadrant}()}
}
\concept{geo}
