% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toptail.R
\name{buff_geo}
\alias{buff_geo}
\title{Create a buffer of n metres for non-projected 'geographical' spatial data}
\usage{
buff_geo(shp, width, ...)
}
\arguments{
\item{shp}{A spatial object with a geographic CRS (e.g. WGS84)
around which a buffer should be drawn}

\item{width}{The distance (in metres) of the buffer (when buffering sp objects)}

\item{...}{Arguments passed to the buffer (see \code{?rgeos::gBuffer} or \code{?sf::st_buffer} for details)}
}
\description{
Solves the problem that buffers will not be circular when used on
non-projected data.
}
\seealso{
Other geo: 
\code{\link{bbox_scale}()},
\code{\link{crs_select_aeq}()},
\code{\link{gclip}()},
\code{\link{geo_bb_matrix}()},
\code{\link{geo_bb}()},
\code{\link{mapshape_available}()},
\code{\link{mapshape}()},
\code{\link{quadrant}()},
\code{\link{reproject}()}
}
\concept{geo}
