% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cents}
\alias{cents}
\alias{cents_sf}
\title{Spatial points representing home locations}
\format{A spatial dataset with 8 rows and 5 variables}
\usage{
data(cents)
}
\description{
These points represent population-weighted centroids of Medium Super Output Area (MSOA) zones within a 1 mile radius of of my home when I was writing this package.
}
\details{
\itemize{
\item geo_code the official code of the zone
\item MSOA11NM name zone name
\item percent_fem the percent female
\item avslope average gradient of the zone
}

Cents was generated from the data repository pct-data: https://github.com/npct/pct-data. This data was accessed from within the pct repo: https://github.com/npct/pct, using the following code:
}
\examples{
\dontrun{
cents
plot(cents)
}

}
\keyword{datasets}
