% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{destinations_sf}
\alias{destinations_sf}
\title{Example destinations data}
\format{
A spatial dataset with 87 features
}
\description{
This dataset represents trip destinations on a different geographic
level than the origins stored in the object \code{cents_sf}.
}
\examples{
destinations_sf
}
\seealso{
Other data: 
\code{\link{cents_sf}},
\code{\link{flow_dests}},
\code{\link{flowlines_sf}},
\code{\link{flow}},
\code{\link{od_data_lines}},
\code{\link{od_data_routes}},
\code{\link{od_data_sample}},
\code{\link{osm_net_example}},
\code{\link{read_table_builder}()},
\code{\link{route_network_sf}},
\code{\link{routes_fast_sf}},
\code{\link{routes_slow_sf}},
\code{\link{zones_sf}}
}
\concept{data}
\keyword{datasets}
