% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google-functions.R
\name{dist_google}
\alias{dist_google}
\title{Return travel network distances and time using the Google Maps API}
\usage{
dist_google(from, to, google_api = "", g_units = "metric")
}
\arguments{
\item{from}{Text string or coordinates (a numeric vector of
\code{length = 2} representing latitude and longitude) representing a point
on Earth.}

\item{to}{Text string or coordinates (a numeric vector of
\code{length = 2} representing latitude and longitude) representing a point
on Earth. This represents the destination of the trip.}

\item{google_api}{String value containing the Google API key to use.}

\item{g_units}{Text string, either metric (default) or imperial.}
}
\description{
Return travel network distances and time using the Google Maps API
}
\section{Details}{

Estimate travel times accounting for the road network - see \url{https://developers.google.com/maps/documentation/distance-matrix/}
Note: Currently returns the json object returned by the Google Maps API and uses the same origins and destinations.
}
\examples{
\dontrun{
 dist_google(from = c(0, 52), to = c(0, 53), google_api = Sys.getenv("GOOGLEDIST"))
}
}

