% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flow_dests}
\alias{flow_dests}
\title{data frame of invented
commuter flows with destinations in a different layer than the origins}
\format{A data frame with 49 rows and 15 columns}
\usage{
data(flow_dests)
}
\description{
data frame of invented
commuter flows with destinations in a different layer than the origins
}
\examples{
\dontrun{
# This is how the dataset was constructed
flow_dests <- flow
flow_dests$Area.of.workplace <- sample(x = destinations$WZ11CD, size = nrow(flow))
flow_dests <- dplyr::rename(flow_dests, WZ11CD = Area.of.workplace)
devtools::use_data(flow_dests)
}

}
\seealso{
Other example data: 
\code{\link{destination_zones}},
\code{\link{flowlines}},
\code{\link{flow}},
\code{\link{route_network}},
\code{\link{routes_fast}},
\code{\link{routes_slow}}
}
\concept{example data}
\keyword{datasets}
