% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{flowlines}
\alias{flowlines}
\title{SpatialLinesDataFrame of commuter flows}
\format{A SpatialLinesDataFrame 49 rows and 15 columns}
\usage{
data(flowlines)
}
\description{
Flow data after conversion to a spatial format
with \code{\link{od2line}} (see \code{\link{flow}}).
}
\keyword{datasets}
