% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{flowlines}
\alias{flowlines}
\alias{flowlines_sf}
\title{spatial lines dataset of commuter flows}
\format{A spatial lines dataset with 49 rows and 15 columns}
\description{
spatial lines dataset of commuter flows
}
\concept{example travel data

Flow data after conversion to a spatial format
with [od2line()] (see [flow()]).}
\keyword{datasets}
