% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flowlines}
\alias{flowlines}
\alias{flowlines_sf}
\title{spatial lines dataset of commuter flows}
\format{A spatial lines dataset with 49 rows and 15 columns}
\description{
Flow data after conversion to a spatial format
with \code{\link[=od2line]{od2line()}} (see \code{\link[=flow]{flow()}}).
}
\seealso{
Other example data: 
\code{\link{destination_zones}},
\code{\link{flow_dests}},
\code{\link{flow}},
\code{\link{route_network}},
\code{\link{routes_fast}},
\code{\link{routes_slow}}
}
\concept{example data}
\keyword{datasets}
