% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flowlines_sf}
\alias{flowlines_sf}
\title{Spatial lines dataset of commuter flows}
\format{
A spatial lines dataset with 49 rows and 15 columns
}
\description{
Flow data after conversion to a spatial format..
}
\seealso{
Other example data: 
\code{\link{destinations_sf}},
\code{\link{flow_dests}},
\code{\link{flow}},
\code{\link{route_network_sf}},
\code{\link{routes_fast_sf}},
\code{\link{routes_slow_sf}}
}
\concept{example data}
\keyword{datasets}
