% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo-functions.R
\name{geo_bb}
\alias{geo_bb}
\alias{bb2poly}
\title{Convert a bounding box to a SpatialPolygonsDataFrame}
\usage{
geo_bb(bb)
}
\arguments{
\item{bb}{the bounding box or spatial object that will be used to crop \code{shp}}
}
\description{
Takes a bounding box as an input and outputs a box in the form of a polygon
}
\examples{
geo_bb(routes_fast)
geo_bb(sp::bbox(routes_fast))
# Simple features implementation:
bb_sf1 <- geo_bb(routes_fast_sf) # keeps CRS
plot(bb_sf1)
plot(routes_fast, add = TRUE)
geo_bb(sf::st_bbox(routes_fast_sf)) # no CRS in bbox
}
