% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo-functions.R
\name{geo_bb_matrix}
\alias{geo_bb_matrix}
\title{Create matrix representing the spatial bounds of an object}
\usage{
geo_bb_matrix(shp)
}
\description{
Converts a range of spatial data formats into a matrix representing the bounding box
}
\seealso{
Other geo: 
\code{\link{bbox_scale}()},
\code{\link{mapshape_available}()},
\code{\link{mapshape}()}
}
\concept{geo}
