% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_projected.R
\name{geo_length}
\alias{geo_length}
\title{Calculate line length of line with geographic or projected CRS}
\usage{
geo_length(shp)
}
\arguments{
\item{shp}{A spatial line object}
}
\description{
Takes a line (represented in sf or sp classes)
and returns a numeric value representing distance in meters.
}
\examples{
geo_length(routes_fast)
geo_length(routes_fast_sf)
}
