% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_projected.R
\name{geo_projected}
\alias{geo_projected}
\title{Perform GIS functions on a temporary, projected version of a spatial object}
\usage{
geo_projected(x, ...)
}
\arguments{
\item{shp}{A spatial object with a geographic (WGS84) coordinate system}

\item{crs}{An optional coordinate reference system (if not provided it is set
automatically by \code{\link{crs_select_aeq}}).}
}
\description{
This function performs operations on projected data.
}
\examples{
shp = st_sf(st_sfc(st_point(c(1, 0))))
geo_projected(shp, st_buffer, dist = 100)
geo_projected(routes_fast, fun = rgeos::gLength, byid = TRUE)
}
