% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gFlow.R
\name{gLines2CyclePath}
\alias{gLines2CyclePath}
\title{Convert straight SpatialLinesDataFrame from flow data into routes in the UK
using the CycleStreets.net API.}
\usage{
gLines2CyclePath(l, plan = "fastest")
}
\arguments{
\item{l}{A SpatialLinesDataFrame object composed of straight lines. l may be
created using the \code{\link{gFlow2line}} function.}

\item{plan}{A text string. Must be either "balanced", "fastest" (default)
or "quietest"}
}
\description{
Convert straight SpatialLinesDataFrame from flow data into routes in the UK
using the CycleStreets.net API.
}
\section{Details}{


You will need to have a cyclestreets api key for this function to work, e.g.:

\code{
mytoken <- readLines("~/Dropbox/dotfiles/cyclestreets-api-key-rl")
}

\code{
Sys.setenv(CYCLESTREET = mytoken) # see http://www.cyclestreets.net/api/
}

\code{
cckey <- Sys.getenv('CYCLESTREET')
}
}
\examples{
library(rgdal)
data(flowlines) # load demo flowlines dataset
flowlines_84 <- sp::spTransform(flowlines, CRS("+init=epsg:4326"))
plot(flowlines_84)

\dontrun{
# set cyclestreets api key: see http://www.cyclestreets.net/api/
cckey <- "f3fe3d078ac34737" # example key (not real)
# plot save the routes
routes_fast <- gLines2CyclePath(flowlines_84)
routes_slow <- gLines2CyclePath(flowlines_84, "quietest")
}
data(routes_fast, routes_slow) # load routes
lines(routes_fast, col = "red")
lines(routes_slow, col = "green")
# Plot for a single line to compare 'fastest' and 'quietest' route
n = 19
plot(flowlines_84[n,])
lines(routes_fast[n,], col = "red")
lines(routes_slow[n,], col = "green")
}

