% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gFlow.R
\name{gLines2CyclePath}
\alias{gLines2CyclePath}
\title{Convert straight SpatialLinesDataFrame from flow data into routes in the UK
using the CycleStreets.net API.}
\usage{
gLines2CyclePath(l, plan = "fastest")
}
\arguments{
\item{l}{A SpatialLinesDataFrame object composed of straight lines. l may be
created using the \code{\link{gFlow2line}} function.}

\item{plan}{A text string. Must be either "balanced", "fastest" (default)
or "quietest"}
}
\description{
Convert straight SpatialLinesDataFrame from flow data into routes in the UK
using the CycleStreets.net API.
}
\section{Details}{


You will need to have a cyclestreets api key for this function to work, e.g.:

\code{
mytoken <- readLines("~/Dropbox/dotfiles/cyclestreets-api-key-rl")
}

\code{
Sys.setenv(CYCLESTREET = mytoken) # see http://www.cyclestreets.net/api/
}

\code{
cckey <- Sys.getenv('CYCLESTREET')
}
}
\examples{
v = data(package = "stplanr")
data(list = v$results[,3])
plot(l)

\dontrun{
Sys.setenv(CYCLESTREET = mytoken) # see http://www.cyclestreets.net/api/
cckey <- Sys.getenv('CYCLESTREET')
routes_fast <- gLines2CyclePath(l)
routes_slow <- gLines2CyclePath(l, "quietest")
}
plot(routes_fast, col = "red", add = TRUE) # previously saved from l
plot(routes_slow, col = "green", add = TRUE)
}

