% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/geo-functions.R
\name{gMapshape}
\alias{gMapshape}
\title{Simplify geometry file of a shapfile.}
\usage{
gMapshape(dsn, percent)
}
\arguments{
\item{dsn}{A character string providing the absolute path to the shapefile to simplify.}

\item{y}{A number between 0 and 100 stating how aggressively to simplify the object. Roughly corresponds with the % space saving of the simplified shapefile compared with the original.}
}
\description{
This is a wrapper funtion for the JavaScript function mapshaper, which must be installed for gMapshaper to work. Writes new file to disk. Thanks to Richard and Adrian Ellison for demonstrating this in R.
}
\examples{
\dontrun{
gMapshape("~/geodata/myShapefile.shp", 5)
}
}

