% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{l_poly}
\alias{l_poly}
\title{Line polygon}
\format{A SpatialPolygon}
\usage{
data(l_poly)
}
\description{
This dataset represents road width for testing.
}
\examples{
\dontrun{
l <- routes_fast[13, ]
l_poly <- geo_projected(l, rgeos::gBuffer, 8)
plot(l_poly)
plot(routes_fast, add = TRUE)
# allocate road width to relevant line
devtools::use_data(l_poly)
}
}
\keyword{datasets}
