% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/od-funs.R
\name{line2df}
\alias{line2df}
\title{Convert geographic line objects to a data.frame with from and to coords}
\usage{
line2df(l)
}
\arguments{
\item{l}{A spatial lines object}
}
\description{
This function returns a data frame with fx and fy and tx and ty variables
representing the beginning and end points of spatial line features respectively.
}
\seealso{
Other lines: 
\code{\link{geo_toptail}()},
\code{\link{line2points}()},
\code{\link{mats2line}()},
\code{\link{n_sample_length}()},
\code{\link{onewaygeo}()},
\code{\link{onewayid}()},
\code{\link{points2line}()},
\code{\link{toptailgs}()},
\code{\link{update_line_geometry}()}
}
\concept{lines}
