% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/od-funs.R
\name{line2points}
\alias{line2points}
\title{Convert a SpatialLinesDataFrame to points at the origin and destination}
\usage{
line2points(l)
}
\arguments{
\item{l}{A SpatialLinesDataFrame}
}
\description{
Convert a SpatialLinesDataFrame to points at the origin and destination
}
\examples{
data(routes_fast)
lpoints <- line2points(routes_fast[2,]) # for a single line
data(flowlines) # load demo flowlines dataset
lpoints <- line2points(flowlines) # for many lines
plot(lpoints) # note overlapping points

}

