% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linefuns.R
\name{line_midpoint}
\alias{line_midpoint}
\title{Find the mid-point of lines}
\usage{
line_midpoint(l)
}
\arguments{
\item{l}{A SpatialLinesDataFrame}
}
\description{
This is a wrapper around \code{\link{SpatialLinesMidPoints}} that allows it to find the midpoint
of lines that are not projected, which have a lat/long CRS.
}
\examples{
data(routes_fast)
line_midpoint(routes_fast)
}

