% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overline.R
\name{lineLabels}
\alias{lineLabels}
\title{Label SpatialLinesDataFrame objects}
\usage{
lineLabels(sl, attrib)
}
\arguments{
\item{sl}{A SpatialLinesDataFrame with overlapping elements}

\item{attrib}{A text string corresponding to a named variable in \code{sl}}
}
\description{
This function adds labels to lines plotted using base graphics. Largely
for illustrative purposes, not designed for publication-quality
graphics.
}
\author{
Barry Rowlingson
}
