% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overline.R
\name{lineLabels}
\alias{lineLabels}
\title{Label SpatialLinesDataFrame objects}
\usage{
lineLabels(sl, attrib)
}
\arguments{
\item{sl}{A SpatialLinesDataFrame with overlapping elements}

\item{attrib}{A text string corresponding to a named variable in \code{sl}}
}
\description{
This function adds labels to lines plotted using base graphics. Largely
for illustrative purposes, not designed for publication-quality
graphics.
}
\seealso{
Other rnet: 
\code{\link{SpatialLinesNetwork}},
\code{\link{calc_catchment_sum}()},
\code{\link{calc_catchment}()},
\code{\link{calc_moving_catchment}()},
\code{\link{calc_network_catchment}()},
\code{\link{find_network_nodes}()},
\code{\link{gsection}()},
\code{\link{islines}()},
\code{\link{overline2}()},
\code{\link{overline}()},
\code{\link{plot,SpatialLinesNetwork,ANY-method}},
\code{\link{plot,sfNetwork,ANY-method}},
\code{\link{sln2points}()},
\code{\link{sum_network_links}()},
\code{\link{sum_network_routes}()}
}
\author{
Barry Rowlingson
}
\concept{rnet}
