% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo-functions.R
\name{mapshape_available}
\alias{mapshape_available}
\title{Does the computer have mapshaper available?}
\usage{
mapshape_available()
}
\description{
This helper function for \code{\link{mapshape}}
determines whether or not the JavaScript library
mapshaper is available.
}
\examples{
mapshape_available()
}

