% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{mapshape_available}
\alias{mapshape_available}
\title{Does the computer have mapshaper available?}
\usage{
mapshape_available()
}
\description{
This helper function for \code{\link[=mapshape]{mapshape()}}
determines whether or not the JavaScript library
mapshaper is available.
}
\examples{
\dontrun{
mapshape_available()
}
}
\seealso{
Other geo: 
\code{\link{bbox_scale}()},
\code{\link{gclip}()},
\code{\link{geo_bb_matrix}()},
\code{\link{geo_bb}()},
\code{\link{mapshape}()},
\code{\link{quadrant}()},
\code{\link{reproject}()}
}
\concept{geo}
