% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linefuns.R
\name{n_vertices}
\alias{n_vertices}
\title{Retrieve the number of vertices from a SpatialLines or SpatialPolygons object}
\usage{
n_vertices(l)
}
\arguments{
\item{l}{A SpatialLines or SpatalPolygons object}
}
\description{
Returns a vector of the same length as the number of lines,
with the number of vertices per line or polygon.
}
\details{
See \url{http://gis.stackexchange.com/questions/58147/} for more information.
}
\examples{
n_vertices(routes_fast)
n_vertices(routes_fast_sf)
}
