% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linefuns.R
\name{n_vertices}
\alias{n_vertices}
\title{Retrieve the number of vertices from a SpatialLines or SpatialPolygons object}
\usage{
n_vertices(l)
}
\arguments{
\item{l}{A SpatialLines or SpatalPolygons object}
}
\description{
Returns a vector of the same length as the number of lines,
with the number of vertices per line or polygon.
}
\details{
See \url{http://gis.stackexchange.com/questions/58147/} for more information.
}
\examples{
n_vertices(routes_fast)
n_vertices(routes_fast_sf)
}
\seealso{
Other lines: 
\code{\link{angle_diff}()},
\code{\link{geo_toptail}()},
\code{\link{is_linepoint}()},
\code{\link{line2df}()},
\code{\link{line2points}()},
\code{\link{line_bearing}()},
\code{\link{line_match}()},
\code{\link{line_midpoint}()},
\code{\link{line_sample}()},
\code{\link{line_segment}()},
\code{\link{line_via}()},
\code{\link{mats2line}()},
\code{\link{n_sample_length}()},
\code{\link{onewaygeo}()},
\code{\link{onewayid}()},
\code{\link{points2line}()},
\code{\link{toptail_buff}()},
\code{\link{toptailgs}()},
\code{\link{update_line_geometry}()}
}
\concept{lines}
