% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/od-funs.R
\name{od_aggregate_from}
\alias{od_aggregate_from}
\title{Summary statistics of trips originating from zones in OD data}
\usage{
od_aggregate_from(flow, attrib = NULL, FUN = sum, ..., col = 1)
}
\arguments{
\item{flow}{A data frame representing origin-destination data.
The first two columns of this data frame should correspond
to the first column of the data in the zones. Thus in \code{\link[=cents_sf]{cents_sf()}},
the first column is geo_code. This corresponds to the first two columns
of \code{\link[=flow]{flow()}}.}

\item{attrib}{character, column names in sl to be aggregated}

\item{FUN}{A function to summarise OD data by}

\item{...}{Additional arguments passed to \code{FUN}}

\item{col}{The column that the OD dataset is grouped by
(1 by default, the first column usually represents the origin)}
}
\description{
This function takes a data frame of OD data and
returns a data frame reporting summary statistics for each unique zone of origin.
}
\details{
It has some default settings: the default summary statistic is \code{sum()} and the
first column in the OD data is assumed to represent the zone of origin.
By default, if \code{attrib} is not set, it summarises all numeric columns.
}
\examples{
od_aggregate_from(flow)
}
\seealso{
Other od: 
\code{\link{od2line}()},
\code{\link{od2odf}()},
\code{\link{od_aggregate_to}()},
\code{\link{od_coords2line}()},
\code{\link{od_coords}()},
\code{\link{od_id}},
\code{\link{od_oneway}()},
\code{\link{od_to_odmatrix}()},
\code{\link{odmatrix_to_od}()},
\code{\link{points2flow}()},
\code{\link{points2odf}()}
}
\concept{od}
