% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{od_data_lines}
\alias{od_data_lines}
\title{Example of desire line representations of origin-destination data from UK Census}
\format{
A data frame (tibble) object
}
\description{
Derived from \code{od_data_sample} showing movement between points represented in \code{cents_sf}
}
\examples{
od_data_lines
}
\keyword{datasets}
