% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{od_data_routes}
\alias{od_data_routes}
\title{Example segment-level route data}
\format{
A data frame (tibble) object
}
\description{
See \code{data-raw/generate-data.Rmd} for details on how this was created.
The dataset shows routes between origins and destinations represented in
\code{od_data_lines}
}
\examples{
od_data_routes
}
\keyword{datasets}
