% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{osm_net_example}
\alias{osm_net_example}
\title{Example of OpenStreetMap road network}
\format{
An sf object
}
\description{
Example of OpenStreetMap road network
}
\examples{
osm_net_example
}
\seealso{
Other data: 
\code{\link{cents_sf}},
\code{\link{destinations_sf}},
\code{\link{flow}},
\code{\link{flow_dests}},
\code{\link{flowlines_sf}},
\code{\link{od_data_lines}},
\code{\link{od_data_routes}},
\code{\link{od_data_sample}},
\code{\link{read_table_builder}()},
\code{\link{route_network_sf}},
\code{\link{route_network_small}},
\code{\link{routes_fast_sf}},
\code{\link{routes_slow_sf}},
\code{\link{zones_sf}}
}
\concept{data}
\keyword{datasets}
