% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{osm_net_example}
\alias{osm_net_example}
\title{Example of OpenStreetMap road network}
\format{An sf object}
\description{
Example of OpenStreetMap road network
}
\examples{
osm_net_example
}
\keyword{datasets}
