% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overline_sf.R
\name{overlaps}
\alias{overlaps}
\title{Identify overlapping lines}
\usage{
overlaps(sl, sl2, pattern = "1F1F00102")
}
\arguments{
\item{sl}{A SpatialLinesDataFrame with overlapping elements}

\item{sl2}{A second \code{LINESTRING} object to aggregate}
}
\description{
Identify overlapping lines
}
