% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overline.R
\name{overline2}
\alias{overline2}
\title{Overlay duplicated lines}
\usage{
overline2(x, attrib, ncores = 1, simplify = TRUE,
  regionalise = 10000)
}
\arguments{
\item{x}{SF data frame of linestrings}

\item{attrib}{character, column name to be summed}

\item{ncores}{integer, how many cores to use in parallel processing}

\item{simplify, }{logical, if TRUE group final segments back into lines}

\item{regionalise, }{integer, during simplification regonalisation is used if the number of segments exceeds this value}
}
\description{
Overlay duplicated lines
}
