% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overline_sf.R
\name{overline_sf2}
\alias{overline_sf2}
\title{Combine 2 \code{sf} \code{LINESTRING} objects into non-overlapping segments and
find the aggregate value of selected columns defined by \code{attrib} along shared segments.}
\usage{
overline_sf2(sl, sl2, attrib = "value", fun = sum)
}
\arguments{
\item{sl}{A SpatialLinesDataFrame with overlapping elements}

\item{sl2}{A second \code{LINESTRING} object to aggregate}

\item{attrib}{A character vector corresponding to the variables in
\code{sl$} on which the function(s) will operate.}

\item{fun}{The function(s) used to aggregate the grouped values (default: sum).
If length of \code{fun} is smaller than \code{attrib} then the functions are
repeated for subsequent attributes.}
}
\description{
Combine 2 \code{sf} \code{LINESTRING} objects into non-overlapping segments and
find the aggregate value of selected columns defined by \code{attrib} along shared segments.
}
\examples{
routes_fast_sf$value = 1
sl = routes_fast_sf[2, ]
sl2 = routes_fast_sf[3, ]
rnet = overline_sf2(sl, sl2, attrib = "value")
plot(rnet, lwd = rnet$value)
rnet2 = overline_sf2(sl, sl2, attrib = c("value", "length"))
plot(rnet2)
sl <- routes_fast_sf[2:5, ]
rnet3 = overline_sf(sl = sl, attrib = c("value", "length"))
rnet4 = overline_sf2(rnet3[1, ], routes_fast_sf[6, ], "value")
plot(rnet3$geometry, lwd = rnet3$value)
plot(rnet4, add = T, col = "red")
}
