% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialLinesNetwork.R
\docType{methods}
\name{plot,sfNetwork,ANY-method}
\alias{plot,sfNetwork,ANY-method}
\title{Plot an sfNetwork}
\usage{
\S4method{plot}{sfNetwork,ANY}(x, component = "sl", ...)
}
\arguments{
\item{x}{The sfNetwork to plot}

\item{component}{The component of the network to plot. Valid values are "sl"
for the geographic (sf) representation or "graph" for the graph
representation.}

\item{...}{Arguments to pass to relevant plot function.}
}
\description{
Plot an sfNetwork
}
\examples{
SLN_sf <- SpatialLinesNetwork(route_network_sf)
plot(SLN_sf)
}
