% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialLinesNetwork.R
\name{plot,sfNetwork,ANY-method}
\alias{plot,sfNetwork,ANY-method}
\title{Plot an sfNetwork}
\usage{
\S4method{plot}{sfNetwork,ANY}(x, component = "sl", ...)
}
\arguments{
\item{x}{The sfNetwork to plot}

\item{component}{The component of the network to plot. Valid values are "sl"
for the geographic (sf) representation or "graph" for the graph
representation.}

\item{...}{Arguments to pass to relevant plot function.}
}
\description{
Plot an sfNetwork
}
\examples{
sln_sf <- SpatialLinesNetwork(route_network_sf)
plot(sln_sf)
}
\seealso{
Other rnet: 
\code{\link{SpatialLinesNetwork}},
\code{\link{calc_catchment_sum}()},
\code{\link{calc_catchment}()},
\code{\link{calc_moving_catchment}()},
\code{\link{calc_network_catchment}()},
\code{\link{find_network_nodes}()},
\code{\link{gsection}()},
\code{\link{islines}()},
\code{\link{lineLabels}()},
\code{\link{overline2}()},
\code{\link{overline}()},
\code{\link{plot,SpatialLinesNetwork,ANY-method}},
\code{\link{sln2points}()},
\code{\link{sum_network_links}()},
\code{\link{sum_network_routes}()}
}
\concept{rnet}
