% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialLinesNetwork.R
\name{plot,SpatialLinesNetwork,ANY-method}
\alias{plot,SpatialLinesNetwork,ANY-method}
\title{Plot a SpatialLinesNetwork}
\usage{
\S4method{plot}{SpatialLinesNetwork,ANY}(x, component = "sl", ...)
}
\arguments{
\item{x}{The SpatialLinesNetwork to plot}

\item{component}{The component of the network to plot. Valid values are "sl"
for the geographic (SpatialLines) representation or "graph" for the graph
representation.}

\item{...}{Arguments to pass to relevant plot function.}
}
\description{
Plot a SpatialLinesNetwork
}
\examples{
sln <- SpatialLinesNetwork(route_network)
plot(sln)
plot(sln, component = "graph")
}
\seealso{
Other rnet: 
\code{\link{SpatialLinesNetwork}},
\code{\link{calc_catchment_sum}()},
\code{\link{calc_catchment}()},
\code{\link{calc_moving_catchment}()},
\code{\link{calc_network_catchment}()},
\code{\link{find_network_nodes}()},
\code{\link{gsection}()},
\code{\link{islines}()},
\code{\link{lineLabels}()},
\code{\link{overline2}()},
\code{\link{overline}()},
\code{\link{plot,sfNetwork,ANY-method}},
\code{\link{sln2points}()},
\code{\link{sum_network_links}()},
\code{\link{sum_network_routes}()}
}
\concept{rnet}
