% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/od-funs.R
\name{points2line}
\alias{points2line}
\title{Convert a series of points, or a matrix of coordinates, into a line}
\usage{
points2line(p)
}
\arguments{
\item{p}{A SpatialPoints obect or matrix representing the coordinates of points.}
}
\description{
This is a simple wrapper around \code{\link{spLines}} that makes the creation of
\code{SpatialLines} objects easy and intuitive
}
\examples{
p = matrix(1:4, ncol = 2)
l = points2line(p)
plot(l)
l = points2line(cents)
plot(l)
p = line2points(routes_fast)
l = points2line(p)
plot(l)
}
