% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/od-funs.R
\name{points2line}
\alias{points2line}
\title{Convert a series of points, or a matrix of coordinates, into a line}
\usage{
points2line(p)
}
\arguments{
\item{p}{A spatial (points) obect or matrix representing the coordinates of points.}
}
\description{
This function makes that makes the creation of \code{sf}
objects with LINESTRING geometries easy.
}
\examples{
l_sf <- points2line(cents_sf)
plot(l_sf)
}
\seealso{
Other lines: 
\code{\link{angle_diff}()},
\code{\link{geo_toptail}()},
\code{\link{is_linepoint}()},
\code{\link{line2df}()},
\code{\link{line2points}()},
\code{\link{line_bearing}()},
\code{\link{line_breakup}()},
\code{\link{line_midpoint}()},
\code{\link{line_segment}()},
\code{\link{line_via}()},
\code{\link{mats2line}()},
\code{\link{n_vertices}()},
\code{\link{onewaygeo}()},
\code{\link{toptail_buff}()}
}
\concept{lines}
