% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnet_boundary_points.R
\name{rnet_boundary_points}
\alias{rnet_boundary_points}
\title{Get points at the beginner and end of linestrings}
\usage{
rnet_boundary_points(rnet)
}
\arguments{
\item{rnet}{An sf object with LINESTRING geometry representing a route
network.}
}
\description{
Get points at the beginner and end of linestrings
}
\examples{
has_sfheaders <- requireNamespace("sfheaders", quietly = TRUE)
if(has_sfheaders) {
rnet <- rnet_roundabout
j <- rnet_boundary_points(rnet)
summary(duplicated(j))
plot(rnet$geometry)
plot(j, add = TRUE)
}
}
