% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnet_group.R
\name{rnet_group}
\alias{rnet_group}
\title{Assign segments in a route network to groups}
\usage{
rnet_group(rnet, cluster_fun = igraph::clusters, d = NULL, char = TRUE)
}
\arguments{
\item{rnet}{An sf object with LINESTRING geometry representing a route
network.}

\item{cluster_fun}{The clustering function to use. Various clustering functions
are available in the \code{igraph} package. Default: \code{\link[igraph:components]{igraph::clusters()}}.}

\item{d}{Optional distance variable used to classify segments that are
close (within a certain distance specified by \code{d}) to each other but not
necessarily touching}

\item{char}{Should the function return a character representation of the groups
(\code{"1"} instead of the numeric value \code{1})? \code{TRUE} by default}
}
\description{
Assign segments in a route network to groups
}
\examples{
rnet <- stplanr::osm_net_example
rnet$group <- rnet_group(rnet)
plot(rnet["group"])
# mapview::mapview(rnet["group"])
rnet$group_25m = rnet_group(rnet, d = 25)
plot(rnet["group_25m"])
rnet$group_walktrap = rnet_group(rnet, igraph::cluster_walktrap)
plot(rnet["group_walktrap"])
rnet$group_louvain = rnet_group(rnet, igraph::cluster_louvain)
plot(rnet["group_louvain"])
rnet$group_fast_greedy = rnet_group(rnet, igraph::cluster_fast_greedy)
plot(rnet["group_fast_greedy"])
}
\concept{}
