% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnet_join.R
\name{rnet_join}
\alias{rnet_join}
\title{Spatial join function that is designed to add columns to a 'target' route network
from a 'source' route network that contains the base geometry, e.g. from OSM}
\usage{
rnet_join(x, y)
}
\description{
Spatial join function that is designed to add columns to a 'target' route network
from a 'source' route network that contains the base geometry, e.g. from OSM
}
