% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{route_network}
\alias{route_network}
\title{SpatialLinesDataFrame representing a route network}
\format{A SpatialLinesDataFrame 80 rows and 1 column}
\usage{
data(route_network)
}
\description{
SpatialLinesDataFrame representing a route network
}
\examples{
\dontrun{
# Generate route network
route_network = overline(routes_fast, "All", fun = sum)
}
}
\keyword{datasets}
