% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{route_network}
\alias{route_network}
\alias{route_network_sf}
\title{spatial lines dataset representing a route network}
\format{A spatial lines dataset 80 rows and 1 column}
\usage{
data(route_network)
}
\description{
The flow of commuters using different segments of the road network represented in the
\code{\link[=flowlines]{flowlines()}} and \code{\link[=routes_fast]{routes_fast()}} datasets
}
\examples{
\dontrun{
# Generate route network
route_network <- overline(routes_fast, "All", fun = sum)
route_network_sf <- sf::st_as_sf(route_network)
}
}
\seealso{
Other example data: 
\code{\link{destination_zones}},
\code{\link{flow_dests}},
\code{\link{flowlines}},
\code{\link{flow}},
\code{\link{routes_fast}},
\code{\link{routes_slow}}
}
\concept{example data}
\keyword{datasets}
