% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{route_network_sf}
\alias{route_network_sf}
\title{spatial lines dataset representing a route network}
\format{
A spatial lines dataset 80 rows and 1 column
}
\description{
The flow of commuters using different segments of the road network represented in the
\code{\link[=flowlines_sf]{flowlines_sf()}} and \code{\link[=routes_fast_sf]{routes_fast_sf()}} datasets
}
\seealso{
Other example data: 
\code{\link{destinations_sf}},
\code{\link{flow_dests}},
\code{\link{flowlines_sf}},
\code{\link{flow}},
\code{\link{routes_fast_sf}},
\code{\link{routes_slow_sf}}
}
\concept{example data}
\keyword{datasets}
