% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{routes_fast}
\alias{routes_fast}
\title{SpatialLinesDataFrame of commuter flows on the travel network}
\format{A SpatialLinesDataFrame 49 rows and 15 columns}
\usage{
data(routes_fast)
}
\description{
Simulated travel route allocated to the transport network
representing the 'fastest' between \code{\link{cents}}
objects
with \code{\link{od2line}} (see \code{\link{flow}}).
}
\keyword{datasets}

