% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{routes_fast}
\alias{routes_fast}
\alias{routes_fast_sf}
\title{spatial lines dataset of commuter flows on the travel network}
\format{A spatial lines dataset with 49 rows and 15 columns}
\usage{
data(routes_fast)
}
\description{
spatial lines dataset of commuter flows on the travel network
}
\keyword{datasets}
