% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{routes_fast}
\alias{routes_fast}
\alias{routes_fast_sf}
\title{spatial lines dataset of commuter flows on the travel network}
\format{A spatial lines dataset with 49 rows and 15 columns}
\usage{
data(routes_fast)
}
\description{
Simulated travel route allocated to the transport network
representing the 'fastest' between \code{\link[=cents]{cents()}}
objects
with \code{\link[=od2line]{od2line()}} (see \code{\link[=flow]{flow()}}).
}
\seealso{
Other example data: 
\code{\link{destination_zones}},
\code{\link{flow_dests}},
\code{\link{flowlines}},
\code{\link{flow}},
\code{\link{route_network}},
\code{\link{routes_slow}}
}
\concept{example data}
\keyword{datasets}
