% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{routes_fast_sf}
\alias{routes_fast_sf}
\title{Spatial lines dataset of commuter flows on the travel network}
\format{
A spatial lines dataset with 42 rows and 15 columns
}
\usage{
routes_fast_sf
}
\description{
Simulated travel route allocated to the transport network
representing the 'fastest' between \code{cents_sf}
objects.
}
\seealso{
Other data: 
\code{\link{cents_sf}},
\code{\link{destinations_sf}},
\code{\link{flow_dests}},
\code{\link{flowlines_sf}},
\code{\link{flow}},
\code{\link{od_data_lines}},
\code{\link{od_data_routes}},
\code{\link{od_data_sample}},
\code{\link{osm_net_example}},
\code{\link{read_table_builder}()},
\code{\link{route_network_sf}},
\code{\link{route_network_small}},
\code{\link{routes_slow_sf}},
\code{\link{zones_sf}}
}
\concept{data}
\keyword{datasets}
