% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{routes_slow}
\alias{routes_slow}
\title{SpatialLinesDataFrame of commuter flows on the travel network}
\format{A SpatialLinesDataFrame 49 rows and 15 columns}
\usage{
data(routes_slow)
}
\description{
SpatialLinesDataFrame of commuter flows on the travel network
}
\keyword{datasets}
