% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{routes_slow}
\alias{routes_slow}
\alias{routes_slow_sf}
\title{spatial lines dataset of commuter flows on the travel network}
\format{A spatial lines dataset 49 rows and 15 columns}
\usage{
data(routes_slow)
}
\description{
spatial lines dataset of commuter flows on the travel network
}
\concept{example travel data

Simulated travel route allocated to the transport network
representing the 'quietest' between [cents()]
objects
with [od2line()] (see [flow()]).}
\keyword{datasets}
