% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stplanr-package.R
\docType{package}
\name{stplanr-package}
\alias{stplanr-package}
\alias{stplanr}
\title{\strong{stplanr: Sustainable Transport Planning with R}}
\description{
The stplanr package provides functions to access
and analyse data for transportation research, including origin-destination analysis,
route allocation and modelling travel patterns.
}
\seealso{
\url{https://github.com/ropensci/stplanr}
}
\author{
Robin Lovelace \email{rob00x@gmail.com}
}
\keyword{package}
