% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stplanr-package.R
\docType{package}
\name{stplanr-package}
\alias{stplanr}
\alias{stplanr-package}
\title{\strong{stplanr: Sustainable Transport Planning with R}}
\description{
The stplanr package provides functions to access
and analyse data for transportation research, including origin-destination analysis,
route allocation and modelling travel patterns.
}
\section{Interesting functions}{

\itemize{
 \item \code{\link{overline}} - Aggregate overlaying route lines and data intelligently
 \item \code{\link{calc_catchment}} - Create a 'catchment area' to show the areas serving a destination
 \item \code{\link{route_cyclestreet}} - Finds the fastest routes for cyclists between two places.
}
}
\author{
Robin Lovelace \email{rob00x@gmail.com}
}
\seealso{
\url{https://github.com/ropensci/stplanr}
}
\keyword{package}

