% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialLinesNetwork.R
\name{summary,SpatialLinesNetwork-method}
\alias{summary,SpatialLinesNetwork-method}
\title{Print a summary of a SpatialLinesNetwork}
\usage{
\S4method{summary}{SpatialLinesNetwork}(object, ...)
}
\arguments{
\item{object}{The SpatialLinesNetwork}

\item{...}{Arguments to pass to relevant summary function.}
}
\description{
Print a summary of a SpatialLinesNetwork
}
\examples{
data(routes_fast)
rnet <- overline(routes_fast, attrib = "length")
sln <- SpatialLinesNetwork(rnet)
summary(sln)
}
