% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osrmfunctions.R
\name{viaroute2sldf}
\alias{viaroute2sldf}
\title{Convert json result of OSRM routing query to SpatialLinesDataFrame}
\usage{
viaroute2sldf(osrmresult)
}
\arguments{
\item{osrmresult}{Single character string or character vector containing
encoded json result(s) of OSRM routing queries.}
}
\description{
Convert json result of OSRM routing query to SpatialLinesDataFrame
}
\section{Details}{

Converts the result of a (successful) OSRM routing query and returns a
SpatialLinesDataFrame containing the route, route summary and instructions.
}

\examples{
\dontrun{
 viaroute2sldf(
   viaroute(startlat = 52.503033,
            startlng = 13.420526,
            endlat = 52.516582,
            endlng = 13.429290)
 )
}
}
