% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo-functions.R
\name{writeGeoJSON}
\alias{writeGeoJSON}
\title{Write to geojson easily}
\usage{
writeGeoJSON(shp, filename)
}
\arguments{
\item{shp}{Spatial data object}

\item{filename}{File name of the output geojson}
}
\description{
Provides a user-friendly wrapper for \code{sf::st_write()}. Note,
\code{geojson_write} from the geojsonio package
provides the same functionality \url{https://github.com/ropensci/geojsonio}.
}
