% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geo-functions.R
\name{writeGeoJSON}
\alias{writeGeoJSON}
\title{Write to geojson easily}
\usage{
writeGeoJSON(shp, filename)
}
\arguments{
\item{shp}{The spatial object a to be cropped}

\item{filename}{File name of the output geojson}
}
\description{
Provides a user-friendly wrapper for rgdal::writeOGR(). Note,
\code{geojson_write} from the geojsonio package
provides the same functionality \url{https://github.com/ropensci/geojsonio}.
}

