% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zones}
\alias{zones}
\alias{zones_sf}
\title{Spatial polygons of home locations for flow analysis.}
\description{
These correspond to the \code{\link[=cents]{cents()}} data.
}
\details{
\itemize{
\item geo_code. the official code of the zone
}
}
\examples{
zones
zones_sf
plot(zones_sf)
}
\keyword{datasets}
