% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zones_sf}
\alias{zones_sf}
\title{Spatial polygons of home locations for flow analysis.}
\description{
These correspond to the \code{cents_sf} data.
}
\details{
\itemize{
\item geo_code. the official code of the zone
}
}
\examples{
library(sf)
zones_sf
plot(zones_sf)
}
\seealso{
Other data: 
\code{\link{cents_sf}},
\code{\link{destinations_sf}},
\code{\link{flow_dests}},
\code{\link{flowlines_sf}},
\code{\link{flow}},
\code{\link{od_data_lines}},
\code{\link{od_data_routes}},
\code{\link{od_data_sample}},
\code{\link{osm_net_example}},
\code{\link{read_table_builder}()},
\code{\link{route_network_sf}},
\code{\link{route_network_small}},
\code{\link{routes_fast_sf}},
\code{\link{routes_slow_sf}}
}
\concept{data}
\keyword{datasets}
